### 2020-11-11 #2836 @whedon
[PRE REVIEW]: ivadomed: A Medical Imaging Deep Learning Toolbox
**Submitting author:** @charleygros (<a href="http://orcid.org/0000-0003-4318-0024">Charley Gros</a>)
**Repository:** <a href="https://github.com/ivadomed/ivadomed" target ="_blank">https://github.com/ivadomed/ivadomed</a>
**Version:** 2.5.0
