### 2020-11-11 #2836 @cMadan
@charleygros, do you have any suggestions for reviewers? (See initial comment in this issue.)

It would also be great if you can add DOIs for articles that are currently missing them in the bib. See some suggestions earlier in this issue (but please double check!).
