### 2020-11-12 32bb2e0
contributing.rst: Redirect to contributing pages
