+++ docs/source/contributing.rst
@@ -1,273 +1,8 @@
 Contributing to ivadomed
 ========================
 
-Introduction
-------------
-
-First off, thanks for taking the time to contribute! 🎉
-
-When contributing to this repository, please first discuss the change
-you wish to make by opening a new `Github
-issue <https://github.com/ivadomed/ivadomed/issues>`__.
-
-Contributions relating to content of the Github repository can be
-submitted through Github pull requests (PR).
-
-PR for bug fixes or new features should be based on the ``master``
-branch.
-
-The following Github documentation may be useful:
-
--  See `Using Pull
-   Requests <https://help.github.com/articles/using-pull-requests>`__
-   for more information about Pull Requests.
--  See `Fork A Repo <http://help.github.com/forking/>`__ for an
-   introduction to forking a repository.
--  See `Creating
-   branches <https://help.github.com/articles/creating-and-deleting-branches-within-your-repository/>`__
-   for an introduction on branching within GitHub.
-
-**For external contributors:** Please fork this repository, make the
-desired changes, and open a Pull request to have your code reviewed and
-merged.
-
-**For internal contributor:** You can open a
-branch (see :ref:`contributing:Opening a branch`) directly in this repository. If you don't
-have the rights, contact the team leader.
-
-Opening an issue
-----------------
-
-Issues (bugs, feature requests, or others) can be submitted `on our
-project's issue
-page. <https://github.com/ivadomed/ivadomed/issues>`__
-
-Before Submitting a New Issue
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-Please take a few seconds to search the issue database in case the issue
-has already been raised.
-
-When reporting an issue, make sure your installation has not been
-tempered with (and if you can, update to the latest release, maybe the
-problem was fixed).
-
-Submitting an Issue
-~~~~~~~~~~~~~~~~~~~
-
-Issue Title
-^^^^^^^^^^^
-
-Try to have a self-descriptive, meaningful issue title, summarizing the
-problem you see.
-
-Examples:
-
--  *Crashes during image cropping*
--  *Add a special mode for multi-class segmentation*
-
-Issue Body
-^^^^^^^^^^
-
-**Describe** the issue and mention the IVADO Medical Imaging version and
-OS that you are using.
-
-If you experience an error, copy/paste the Terminal output (include your
-syntax) and please follow these guidelines for clarity:
-
--  If there is less than 10 lines of text, embed it directly in your
-   comment in Github. Use "~~~" to format as code.
--  If there is 10+ lines, either use an external website such as
-   `pastebin <https://pastebin.com/>`__ (copy/paste your text and
-   include the URL in your comment), or use `collapsable Github markdown
-   capabilities <https://gist.github.com/ericclemmons/b146fe5da72ca1f706b2ef72a20ac39d#using-details-in-github>`__.
-
-Add useful information such as screenshots, etc.
-
-If you submit a feature request, provide a *usage scenario*, imagining
-how the feature would be used (ideally inputs, a sequence of commands,
-and a desired outcome). Also provide references to any theoretical work
-to help the reader better understand the feature.
-
-Opening a Branch
-----------------
-
-If you are part of the core developer team, you can open a branch
-directly in this repository. Prefix the branch name with a personal
-identifier and a forward slash; If the branch you are working on is in
-response to an issue, provide the issue number; Add some text that make
-the branch name meaningful.
-
-Examples:
-
--  ``ol/100-fixup-lr-scheduler``
--  ``ab/loader-pep8``
-
-Developing
-----------
-
-Conflicts
-~~~~~~~~~
-
-Make sure the PR changes are not in conflict with the master branch.
-
-Code style
-~~~~~~~~~~
-
[…]
