### 2020-11-12 f6d8b43
contributing.rst: Update DokuWiki (open_source -> software)
