+++ docs/source/contributing.rst
@@ -4,5 +4,5 @@ Contributing to ivadomed
 Thank you for your interest in contributing to ivadomed! This project uses the following pages to guide new contributions:
 
   * The `ivadomed GitHub repository <https://github.com/ivadomed/ivadomed>`_ is where the source code for the project is maintained, and where new contributions are submitted to.
-  * The `NeuroPoly Contributing Guidelines <https://www.neuro.polymtl.ca/open_source/contributing>`_ provide instructions for development workflows, such as reporting issues or submitting pull requests.
+  * The `NeuroPoly Contributing Guidelines <https://www.neuro.polymtl.ca/software/contributing>`_ provide instructions for development workflows, such as reporting issues or submitting pull requests.
   * The `ivadomed Developer Wiki <https://github.com/ivadomed/ivadomed/wiki>`_ acts as a knowledge base for documenting internal design decisions specific to the ivadomed codebase. It also contains step-by-step walkthroughs for common ivadomed maintainer tasks.
