### 2020-11-12 2ac8317
segment_volume: produce one output per class
