### 2020-11-12 4e057eb
command segment: adapt to new segment volume
