### 2020-11-12 8fae08a
object_detection: adapt to new segment_volume
