### 2020-11-12 dba6a59
testing: modify test to new segment_volume
