### 2020-11-12 6aef163
deal with multichannel metadata
