### 2020-11-13 e788c6e
segment command: multichannel
