### 2020-11-13 8c60446
adapt segment volume to multichannel
