### 2020-11-13 a0a7438
pred type to float32
