### 2020-11-15 #2836 @charleygros
Many thanks @cMadan for accepting to be the editor for this submission!

Sorry for the delayed reply.

> @charleygros, do you have any suggestions for reviewers? (See initial comment in this issue.)

Suggestions for reviewers from the provided list:
- rlizzo
- NMontanaBrown
- Bsingstad
- justusschock
- lbugnon
- esadr
- fernandoandreotti

> It would also be great if you can add DOIs for articles that are currently missing them in the bib. See some suggestions earlier in this issue (but please double check!).

Sure, I'm happy do this ASAP and will keep you posted.
