### 2020-11-15 41e8bd4
paper.bib: Wang2019
