+++ paper.bib
@@ -1,16 +1,9 @@
-@article{wang_adaptive_2019,
-	title = {Adaptive {Wing} {Loss} for {Robust} {Face} {Alignment} via {Heatmap} {Regression}},
-	url = {http://arxiv.org/abs/1904.07399},
-	abstract = {Heatmap regression with a deep network has become one of the mainstream approaches to localize facial landmarks. However, the loss function for heatmap regression is rarely studied. In this paper, we analyze the ideal loss function properties for heatmap regression in face alignment problems. Then we propose a novel loss function, named Adaptive Wing loss, that is able to adapt its shape to different types of ground truth heatmap pixels. This adaptability penalizes loss more on foreground pixels while less on background pixels. To address the imbalance between foreground and background pixels, we also propose Weighted Loss Map, which assigns high weights on foreground and difficult background pixels to help training process focus more on pixels that are crucial to landmark localization. To further improve face alignment accuracy, we introduce boundary prediction and CoordConv with boundary coordinates. Extensive experiments on different benchmarks, including COFW, 300W and WFLW, show our approach outperforms the state-of-the-art by a significant margin on various evaluation metrics. Besides, the Adaptive Wing loss also helps other heatmap regression tasks. Code will be made publicly available at https://github.com/protossw512/AdaptiveWingLoss.},
-	urldate = {2019-12-15},
-	journal = {arXiv:1904.07399 [cs]},
-	author = {Wang, Xinyao and Bo, Liefeng and Fuxin, Li},
-	month = oct,
-	year = {2019},
-	note = {arXiv: 1904.07399},
-	keywords = {Computer Science - Computer Vision and Pattern Recognition},
-	annote = {Comment: Camera-ready version for ICCV 2019},
-	file = {arXiv Fulltext PDF:C\:\\Users\\Lucas\\Zotero\\storage\\ZITJDN2F\\Wang et al. - 2019 - Adaptive Wing Loss for Robust Face Alignment via H.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\Lucas\\Zotero\\storage\\9BT2TJFQ\\1904.html:text/html}
+@inproceedings{wang2019adaptive,
+  title={Adaptive wing loss for robust face alignment via heatmap regression},
+  author={Wang, Xinyao and Bo, Liefeng and Fuxin, Li},
+  booktitle={Proceedings of the IEEE International Conference on Computer Vision},
+  pages={6971--6981},
+  year={2019}
 }
 
 
