+++ paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{wang2019adaptive,
+@inproceedings{wang_adaptive_2019,
   title={Adaptive wing loss for robust face alignment via heatmap regression},
   author={Wang, Xinyao and Bo, Liefeng and Fuxin, Li},
   booktitle={Proceedings of the IEEE International Conference on Computer Vision},
