+++ paper.bib
@@ -103,22 +103,15 @@ Publisher: Nature Publishing Group},
 }
 
 
-
 @article{nair_exploring_2018,
-	title = {Exploring {Uncertainty} {Measures} in {Deep} {Networks} for {Multiple} {Sclerosis} {Lesion} {Detection} and {Segmentation}},
-	url = {http://arxiv.org/abs/1808.01200},
-	abstract = {Deep learning (DL) networks have recently been shown to outperform other segmentation methods on various public, medical-image challenge datasets [3,11,16], especially for large pathologies. However, in the context of diseases such as Multiple Sclerosis (MS), monitoring all the focal lesions visible on MRI sequences, even very small ones, is essential for disease staging, prognosis, and evaluating treatment efficacy. Moreover, producing deterministic outputs hinders DL adoption into clinical routines. Uncertainty estimates for the predictions would permit subsequent revision by clinicians. We present the first exploration of multiple uncertainty estimates based on Monte Carlo (MC) dropout [4] in the context of deep networks for lesion detection and segmentation in medical images. Specifically, we develop a 3D MS lesion segmentation CNN, augmented to provide four different voxel-based uncertainty measures based on MC dropout. We train the network on a proprietary, large-scale, multi-site, multi-scanner, clinical MS dataset, and compute lesion-wise uncertainties by accumulating evidence from voxel-wise uncertainties within detected lesions. We analyze the performance of voxel-based segmentation and lesion-level detection by choosing operating points based on the uncertainty. Empirical evidence suggests that uncertainty measures consistently allow us to choose superior operating points compared only using the network's sigmoid output as a probability.},
-	urldate = {2020-09-28},
-	journal = {arXiv:1808.01200 [cs]},
-	author = {Nair, Tanya and Precup, Doina and Arnold, Douglas L. and Arbel, Tal},
-	month = oct,
-	year = {2018},
-	note = {arXiv: 1808.01200},
-	keywords = {Computer Science - Computer Vision and Pattern Recognition},
-	annote = {Comment: Updated references in Introduction; Accepted to the 21st International Conference on Medical Image Computing and Computer Assisted Intervention (MICCAI 2018)},
-	file = {arXiv Fulltext PDF:C\:\\Users\\Lucas\\Zotero\\storage\\PQFS7ZXK\\Nair et al. - 2018 - Exploring Uncertainty Measures in Deep Networks fo.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\Lucas\\Zotero\\storage\\6RDHLL3J\\1808.html:text/html}
+  title={Exploring uncertainty measures in deep networks for multiple sclerosis lesion detection and segmentation},
+  author={Nair, Tanya and Precup, Doina and Arnold, Douglas L and Arbel, Tal},
+  journal={Medical image analysis},
+  volume={59},
+  pages={101557},
+  year={2020},
+  publisher={Elsevier}
 }
-
 
 
 @article{wang_aleatoric_2019,
