### 2020-11-15 99adb62
paper.bib: havaei2016hemis
