+++ paper.bib
@@ -153,14 +153,14 @@ Publisher: Nature Publishing Group},
   year={2017}
 }
 
-@inproceedings{havaei2016hemis,
-  title={Hemis: Hetero-modal image segmentation},
+
+@article{havaei2016hemis,
+  title={HeMIS: Hetero-Modal Image Segmentation},
   author={Havaei, Mohammad and Guizard, Nicolas and Chapados, Nicolas and Bengio, Yoshua},
-  booktitle={International Conference on Medical Image Computing and Computer-Assisted Intervention},
-  pages={469--477},
-  year={2016},
-  organization={Springer}
+  journal={arXiv preprint arXiv:1607.05194},
+  year={2016}
 }
+
 
 @inproceedings{Cohen2017countception,
 abstract = {Counting objects in digital images is a process that should be replaced by machines. This tedious task is time consuming and prone to errors due to fatigue of human annotators. The goal is to have a system that takes as input an image and returns a count of the objects inside and justification for the prediction in the form of object localization. We repose a problem, originally posed by Lempitsky and Zisserman, to instead predict a count map which contains redundant counts based on the receptive field of a smaller regression network. The regression network predicts a count of the objects that exist inside this frame. By processing the image in a fully convolutional way each pixel is going to be accounted for some number of times, the number of windows which include it, which is the size of each window, (i.e., 32x32 = 1024). To recover the true count we take the average over the redundant predictions. Our contribution is redundant counting instead of predicting a density map in order to average over errors. We also propose a novel deep neural network architecture adapted from the Inception family of networks called the Count-ception network. Together our approach results in a 20{\%} relative improvement (2.9 to 2.3 MAE) over the state of the art method by Xie, Noble, and Zisserman in 2016.},
