### 2020-11-15 757b94c
paper.bib: DOI roy_quicknat_2018
