+++ paper.bib
@@ -176,6 +176,7 @@ url = {http://arxiv.org/abs/1703.08710},
 year = {2017}
 }
 
+
 @inproceedings{Huang2017densenet,
 archivePrefix = {arXiv},
 arxivId = {1608.06993},
@@ -228,6 +229,7 @@ year = {2016}
 	abstract = {Whole brain segmentation from structural magnetic resonance imaging (MRI) is a prerequisite for most morphological analyses, but is computationally intense and can therefore delay the availability of image markers after scan acquisition. We introduce QuickNAT, a fully convolutional, densely connected neural network that segments a {\textbackslash}revision\{MRI brain scan\} in 20 seconds. To enable training of the complex network with millions of learnable parameters using limited annotated data, we propose to first pre-train on auxiliary labels created from existing segmentation software. Subsequently, the pre-trained model is fine-tuned on manual labels to rectify errors in auxiliary labels. With this learning strategy, we are able to use large neuroimaging repositories without manual annotations for training. In an extensive set of evaluations on eight datasets that cover a wide age range, pathology, and different scanners, we demonstrate that QuickNAT achieves superior segmentation accuracy and reliability in comparison to state-of-the-art methods, while being orders of magnitude faster. The speed up facilitates processing of large data repositories and supports translation of imaging biomarkers by making them available within seconds for fast clinical decision making.},
 	urldate = {2020-09-28},
 	journal = {arXiv:1801.04161 [cs]},
+	doi = {10.1016/j.neuroimage.2018.11.042},
 	author = {Roy, Abhijit Guha and Conjeti, Sailesh and Navab, Nassir and Wachinger, Christian},
 	month = nov,
 	year = {2018},
@@ -236,6 +238,7 @@ year = {2016}
 	annote = {Comment: Accepted for Publication at NeuroImage},
 	file = {arXiv Fulltext PDF:C\:\\Users\\Lucas\\Zotero\\storage\\864K65TV\\Roy et al. - 2018 - QuickNAT A Fully Convolutional Network for Quick .pdf:application/pdf}
 }
+
 
 @inproceedings{isensee2017brain,
   title={Brain tumor segmentation and radiomics survival prediction: Contribution to the \uppercase{BRATS} 2017 challenge},
