+++ paper.bib
@@ -249,14 +249,17 @@ year = {2016}
   organization={Springer}
 }
 
+
 @inproceedings{milletari2016v,
   title={V-net: Fully convolutional neural networks for volumetric medical image segmentation},
   author={Milletari, Fausto and Navab, Nassir and Ahmadi, Seyed-Ahmad},
   booktitle={2016 fourth international conference on 3D vision (3DV)},
   pages={565--571},
+  doi={10.1109/3DV.2016.79},
   year={2016},
   organization={IEEE}
 }
+
 
 @inproceedings{wong20183d,
   title={3D segmentation with exponential logarithmic loss for highly unbalanced object sizes},
