+++ paper.bib
@@ -3,7 +3,8 @@
   author={Wang, Xinyao and Bo, Liefeng and Fuxin, Li},
   booktitle={Proceedings of the IEEE International Conference on Computer Vision},
   pages={6971--6981},
-  year={2019}
+  year={2019},
+  doi={10.1109/iccv.2019.00707}
 }
 
 
@@ -103,14 +104,15 @@ Publisher: Nature Publishing Group},
 }
 
 
-@article{nair_exploring_2018,
+@article{,
   title={Exploring uncertainty measures in deep networks for multiple sclerosis lesion detection and segmentation},
   author={Nair, Tanya and Precup, Doina and Arnold, Douglas L and Arbel, Tal},
   journal={Medical image analysis},
   volume={59},
   pages={101557},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.media.2019.101557}
 }
 
 
@@ -158,7 +160,8 @@ Publisher: Nature Publishing Group},
   title={HeMIS: Hetero-Modal Image Segmentation},
   author={Havaei, Mohammad and Guizard, Nicolas and Chapados, Nicolas and Bengio, Yoshua},
   journal={arXiv preprint arXiv:1607.05194},
-  year={2016}
+  year={2016},
+  doi={10.1007/978-3-319-46723-8_54}
 }
 
 
@@ -173,7 +176,8 @@ file = {:Users/ieee8023/Library/Application Support/Mendeley Desktop/Downloaded/
 mendeley-groups = {neuralnetworks/counting/core-counting},
 title = {{Count-ception: Counting by Fully Convolutional Redundant Counting}},
 url = {http://arxiv.org/abs/1703.08710},
-year = {2017}
+year = {2017},
+doi = {10.1109/iccvw.2017.9}
 }
 
 
@@ -255,7 +259,7 @@ year = {2016}
   author={Milletari, Fausto and Navab, Nassir and Ahmadi, Seyed-Ahmad},
   booktitle={2016 fourth international conference on 3D vision (3DV)},
   pages={565--571},
-  doi={10.1109/3DV.2016.79},
+  doi={10.1109/3dv.2016.79},
   year={2016},
   organization={IEEE}
 }
@@ -267,7 +271,8 @@ year = {2016}
   booktitle={International Conference on Medical Image Computing and Computer-Assisted Intervention},
   pages={612--619},
   year={2018},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-00931-1_70}
 }
 
 @ARTICLE{Willemink2020-au,
