### 2020-11-15 25e1d28
Fix DOIs in paper.bib (#527)
* paper.md: rm XX ORCID

* paper.bib: Wang2019

* paper.bib: Wang2019 ref name

* paper.bib: Nair2020

* paper.bib: havaei2016hemis

* paper.bib: DOI roy_quicknat_2018

* paper.bib: DOI milletari2016v

* paper.bib: DOIs
