### 2020-11-15 1dbdd42
paper.bib: typo (#528)
