### 2020-11-15 1ffb3f0
paper.md: path overview_title.png
