### 2020-11-17 546f586
adapwingloss: adapt to 3D
