### 2020-11-18 0d673c3
Update README (#518)
* update readme

* fix badge

Co-authored-by: charleygros <charley.gros@gmail.com>
