### 2020-11-18 56952a5
Change path for images (#521)
* change path for images

* change path logo

* update path images

* remove images

Co-authored-by: charleygros <charley.gros@gmail.com>
