+++ README.md
@@ -1,5 +1,5 @@
   
-![ivadomed Overview](https://github.com/ivadomed/ivadomed/raw/master/images/overview_title.png)
+![ivadomed Overview](https://raw.githubusercontent.com/ivadomed/doc-figures/main/index/overview_title.png)
 
 [![Coverage Status](https://coveralls.io/repos/github/ivadomed/ivadomed/badge.svg?branch=master)](https://coveralls.io/github/ivadomed/ivadomed?branch=master)
 ![](https://github.com/neuropoly/ivadomed/workflows/Python%20package/badge.svg)
@@ -12,9 +12,9 @@ The technical documentation is available [here](https://ivadomed.org).
 
 ## Contributors
 <p float="left">
-  <img src="https://github.com/ivadomed/ivadomed/raw/master/images/neuropoly_logo.png" height="80" />
-  <img src="https://github.com/ivadomed/ivadomed/raw/master/images/mila_logo.png" height="80" />
-  <img src="https://github.com/ivadomed/ivadomed/raw/master/images/ivado_logo.png" height="80" />
+  <img src="https://raw.githubusercontent.com/ivadomed/doc-figures/main/contributors/neuropoly_logo.png" height="80" />
+  <img src="https://raw.githubusercontent.com/ivadomed/doc-figures/main/contributors/mila_logo.png" height="80" />
+  <img src="https://raw.githubusercontent.com/ivadomed/doc-figures/main/contributors/ivado_logo.png" height="80" />
 </p>
 
 This project results from a collaboration between the [NeuroPoly Lab](https://www.neuro.polymtl.ca/)
