### 2020-11-18 ecb76e0
Update ivadomed/losses.py
Co-authored-by: charleygros <charley.gros@gmail.com>
