### 2020-11-18 5d2d25f
Update ivadomed/inference.py
Co-authored-by: charleygros <charley.gros@gmail.com>
