### 2020-11-18 9096673
add error message for bad metadata choice
