### 2020-11-18 c905522
Create sample to balance metadata (#503)
* Change redundant title

* change file name to fit title

* models.rst: add sectionning

* correct faulty reference

* tumor type film

* save post-processed prediction

* uncertainty for multilabel

* remove space

* add uncertainty threshold

* threshold according to uncertainty

* update config files

* option if thr is -1

* get data for each image in batch

* add FiLM metadata choice to metadata

* add ohe dict

* encode film data

* remove lumbar split

* update doc string

* update RTD

* change contrast for contrasts

* set data metadata_choice
[…]
