### 2020-11-18 b840602
loader.py: add comment to clarify
