### 2020-11-18 6ecb85d
add warning if missing channel in segment
