### 2020-11-18 3086aab
change segment volume to stop asumption of one output
