### 2020-11-18 8f99591
loader.py: fix type stack_input -> stack_gt
