### 2020-11-19 8f3db0b
Support multiclass if first class missing (#522)
* multiclass: support multiclass if first class missing

* testing: fix if target class 0 is not present

* evaluation: fix index out of error

* loader.py: add comment to clarify
