### 2020-11-20 7edec51
Multiclass and multichannel support for segment volume (#524)
* segment_volume: produce one output per class

* command segment: adapt to new segment volume

* object_detection: adapt to new segment_volume

* testing: modify test to new segment_volume

* support multichannel models

* deal with multichannel metadata

* segment_volume: metadata init

* segment command: multichannel

* adapt segment volume to multichannel

* pred type to float32

* Update ivadomed/inference.py

Co-authored-by: charleygros <charley.gros@gmail.com>

* document new output for segment volume

* add warning if missin
[…]
