+++ testing/test_segment_volume.py
@@ -68,8 +68,8 @@ def test_segment_volume_2d():
     with open(PATH_CONFIG, 'w') as fp:
         json.dump(config, fp)
 
-    nib_img = imed_inference.segment_volume(PATH_MODEL, IMAGE_PATH, ROI_PATH)
-
+    nib_lst, _ = imed_inference.segment_volume(PATH_MODEL, [IMAGE_PATH], ROI_PATH)
+    nib_img = nib_lst[0]
     assert np.squeeze(nib_img.get_fdata()).shape == nib.load(IMAGE_PATH).shape
     assert (nib_img.dataobj.max() <= 1.0) and (nib_img.dataobj.min() >= 0.0)
     assert nib_img.dataobj.dtype == 'float32'
@@ -133,7 +133,8 @@ def test_segment_volume_3d():
     with open(PATH_CONFIG, 'w') as fp:
         json.dump(config, fp)
 
-    nib_img = imed_inference.segment_volume(PATH_MODEL, IMAGE_PATH)
+    nib_lst, _ = imed_inference.segment_volume(PATH_MODEL, [IMAGE_PATH])
+    nib_img = nib_lst[0]
     assert np.squeeze(nib_img.get_fdata()).shape == nib.load(IMAGE_PATH).shape
     assert (nib_img.dataobj.max() <= 1.0) and (nib_img.dataobj.min() >= 0.0)
     assert nib_img.dataobj.dtype == 'float32'
