+++ ivadomed/inference.py
@@ -95,7 +95,7 @@ def pred_to_nib(data_lst, z_lst, fname_ref, fname_out, slice_axis, debug=False, 
     return nib_pred
 
 
-def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
+def segment_volume(folder_model, fname_images, gpu_number=0, options=None):
     """Segment an image.
     Segment an image (`fname_image`) using a pre-trained model (`folder_model`). If provided, a region of interest
     (`fname_roi`) is used to crop the image prior to segment it.
@@ -104,7 +104,8 @@ def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
             (1) the model ('folder_model/folder_model.pt') to use
             (2) its configuration file ('folder_model/folder_model.json') used for the training,
             see https://github.com/neuropoly/ivadomed/wiki/configuration-file
-        fname_image (str): image filename (e.g. .nii.gz) to segment.
+        fname_images (list): list of image filenames (e.g. .nii.gz) to segment. Multichannel models require multiple
+            images to segment, e.i., len(fname_images) > 1.
         gpu_number (int): Number representing gpu number if available.
         options (dict): Contains postprocessing steps and prior filename (fname_prior) which is an image filename
             (e.g., .nii.gz) containing processing information (e.i., spinal cord segmentation, spinal location or MS
@@ -112,7 +113,9 @@ def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
             e.g., spinal cord centerline, used to crop the image prior to segment it if provided.
             The segmentation is not performed on the slices that are empty in this image.
     Returns:
-        nibabelObject: Object containing the soft segmentation.
+        list: List of nibabel objects containing the soft segmentation(s), one per prediction class.
+        list: List of target suffix associated with each prediction in `pred_list`
+
     """
     # Define device
     cuda_available = torch.cuda.is_available()
@@ -162,6 +165,7 @@ def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
         if 'object_detection_params' in context and \
                 context['object_detection_params']['object_detection_path'] is not None:
             imed_obj_detect.bounding_box_prior(fname_prior, metadata, slice_axis)
+            metadata = [metadata] * len(fname_images)
 
     # Compose transforms
     _, _, transform_test_params = imed_transforms.get_subdatasets_transforms(context["transformation"])
@@ -173,7 +177,7 @@ def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
         print("\nWARNING: fname_roi has not been specified, then the entire volume is processed.")
         loader_params["slice_filter_params"]["filter_empty_mask"] = False
 
-    filename_pairs = [([fname_image], None, fname_roi, [metadata])]
+    filename_pairs = [(fname_images, None, fname_roi, metadata if isinstance(metadata, list) else [metadata])]
 
     kernel_3D = bool('Modified3DUNet' in context and context['Modified3DUNet']['applied']) or \
                 not context['default_model']['is_2d']
@@ -242,16 +246,17 @@ def segment_volume(folder_model, fname_image, gpu_number=0, options=None):
             preds = preds.cpu()
 
         # Set datatype to gt since prediction should be processed the same way as gt
-        for modality in batch['input_metadata']:
-            modality[0]['data_type'] = 'gt'
+        for b in batch['input_metadata']:
+            for modality in b:
+                modality['data_type'] = 'gt'
 
         # Reconstruct 3D object
         for i_slice in range(len(preds)):
             if "bounding_box" in batch['input_metadata'][i_slice][0]:
                 imed_obj_detect.adjust_undo_transforms(undo_transforms.transforms, batch, i_slice)
 
[…]
