### 2020-11-20 a16e6e6
automate_training: add new parameter to change multiple params (#533)
