### 2020-11-20 1409ab2
Add option to select subjects according to metadata
