### 2020-11-20 22bfe4e
main.py: subject selection param
