+++ ivadomed/main.py
@@ -98,7 +98,9 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
         train_lst, valid_lst, test_lst = imed_loader_utils.get_subdatasets_subjects_list(context["split_dataset"],
                                                                                          context['loader_parameters']
                                                                                          ['bids_path'],
-                                                                                         log_directory)
+                                                                                         log_directory,
+                                                                                         context["loader_parameters"]
+                                                                                         ['subject_selection'])
 
     # Loader params
     loader_params = copy.deepcopy(context["loader_parameters"])
