### 2020-11-20 bbbfbb5
subject_selection: update to avoid bug when list is empty
