### 2020-11-23 3a9258d
Make is_2d retrocompatibility (#535)
* models.py: fix retrocompatibility issues with is_2d

* testing: forgotten line
