### 2020-11-23 #2836 @arfon
It looks like the Crossref API is timing out. I managed to get this to run locally (the DOIs are fine):

```

OK DOIs

- 10.1109/iccv.2019.00707 is OK
- 10.1016/j.nicl.2018.03.029 is OK
- 10.1016/j.neuroimage.2018.09.081 is OK
- 10.14245/ns.1938396.198 is OK
- 10.1038/sdata.2016.44 is OK
- 10.1016/j.media.2019.101557 is OK
- 10.1016/j.neucom.2019.01.103 is OK
- 10.1007/978-3-319-46723-8_54 is OK
- 10.1109/iccvw.2017.9 is OK
- 10.1109/CVPR.2016.90 is OK
- 10.1016/j.neuroimage.2018.11.042 is OK
- 10.1109/3dv.2016.79 is OK
- 10.1007/978-3-030-00931-1_70 is OK

MISSING DOIs

- None

INVALID DOIs

- None
```
