### 2020-11-24 ef8609e
loader/utils.py: Add create_bids_dataframe function
