### 2020-11-25 624f862
adaptative: add task attribute
