### 2020-11-25 7c10596
Define path with os.join
