### 2020-11-25 f0a8990
Remove unneccesary print
