### 2020-11-25 4e5eaf7
bids_config default to None if not present in config file
