### 2020-11-25 eed1483
Change file filtering to work with relative paths and avoid df concat
