### 2020-11-26 #2868 @whedon
[REVIEW]: ivadomed: A Medical Imaging Deep Learning Toolbox
**Submitting author:** @charleygros (<a href="http://orcid.org/0000-0003-4318-0024">Charley Gros</a>)
**Repository:** <a href="https://github.com/ivadomed/ivadomed" target ="_blank">https://github.com/ivadomed/ivadomed</a>
**Version:** v2.7.1
