### 2020-11-27 c3d1a88
Read participants.tsv only if present
