### 2020-11-27 99534ab
Minor corrections script dev
