### 2020-12-01 1fdc3ea
Add filtering of dataframe by "data_types"
