### 2020-12-01 ad59c66
adaptative loader for multiclass
