### 2020-12-01 776001f
change target_lst to target_contrast
