### 2020-12-01 ad032fc
add documentation on HeMIS params
