+++ docs/source/configuration_file.rst
@@ -323,6 +323,12 @@ HeMISUnet (Optional)
 -  ``missing_probability_growth``: Float. Controls missing probability
    growth at each epoch: at each epoch, the ``missing_probability`` is
    modified with the exponent ``missing_probability_growth``.
+- ``contrasts``: List. List of contrasts used as input.
+- ``ram``. Bool. Indicates whether to save the data in ram or not.
+- ``hdf5_path``: Str. Path to hdf5 files containing the dataset.
+- ``csv_path``: Str. Path to dataframe.
+- ``target_contrast``: Str. Contrast of GT to be used.
+- `` Contrast label used for ROI cropping.``: Str. Contrast of ROI to be used.
 
 .. _Modified3DUNet:
 
