### 2020-12-01 3da39e3
update test with roi_contrast
