### 2020-12-02 c42fb48
Multichannel support for convert_to_onnx script (#544)
