### 2020-12-02 2ed340a
Update docs/source/configuration_file.rst
Co-authored-by: charleygros <charley.gros@gmail.com>
