### 2020-12-02 e9b021e
Update docs/source/configuration_file.rst
Co-authored-by: charleygros <charley.gros@gmail.com>
