+++ docs/source/configuration_file.rst
@@ -326,7 +326,7 @@ HeMISUnet (Optional)
 - ``contrasts``: List. List of contrasts used as input.
 - ``ram``. Bool. Indicates whether to save the data in ram or not.
 - ``hdf5_path``: Str. Path to hdf5 files containing the dataset.
-- ``csv_path``: Str. Path to dataframe.
+- ``csv_path``: Str. Path to dataframe (csv file).
 - ``target_contrast``: Str. Contrast label used for ground truth.
 - `` Contrast label used for ROI cropping.``: Str. Contrast whom the label file is used for ROI cropping.
 
