### 2020-12-02 7632c66
Update docs/source/configuration_file.rst
Co-authored-by: charleygros <charley.gros@gmail.com>
