+++ docs/source/configuration_file.rst
@@ -327,7 +327,7 @@ HeMISUnet (Optional)
 - ``ram``. Bool. Indicates whether to save the data in ram or not.
 - ``hdf5_path``: Str. Path to hdf5 files containing the dataset.
 - ``csv_path``: Str. Path to dataframe (csv file).
-- ``target_contrast``: Str. Contrast label used for ground truth.
+- ``target_contrast``: Str. Contrast whom the label file is used as ground truth.
 - `` Contrast label used for ROI cropping.``: Str. Contrast whom the label file is used for ROI cropping.
 
 .. _Modified3DUNet:
