### 2020-12-02 6de99ea
Update docs/source/configuration_file.rst
Co-authored-by: charleygros <charley.gros@gmail.com>
