### 2020-12-02 09d2505
Update docs/source/configuration_file.rst
