### 2020-12-02 0b6c477
loader.py: add multiple target_filename per class
