+++ ivadomed/loader/loader.py
@@ -781,12 +781,17 @@ class BidsDataset(MRI2DSegmentationDataset):
                     print("Subject without derivative, skipping.")
                     continue
                 derivatives = subject.get_derivatives("labels")
-                target_filename, roi_filename = [None] * len(target_suffix), None
+                target_filename, roi_filename = [[]] * len(target_suffix), None
 
                 for deriv in derivatives:
-                    for idx, suffix in enumerate(target_suffix):
-                        if deriv.endswith(subject.record["modality"] + suffix + ".nii.gz"):
-                            target_filename[idx] = deriv
+                    for idx, suffix_list in enumerate(target_suffix):
+                        # If suffix_list is a string, then only one rater annotation per class is available.
+                        # Otherwise, multiple raters segmented the same class.
+                        if type(suffix_list) == str:
+                            suffix_list = [suffix_list]
+                        for suffix in suffix_list:
+                            if deriv.endswith(subject.record["modality"] + suffix + ".nii.gz"):
+                                target_filename[idx].append(deriv)
 
                     if not (self.roi_params["suffix"] is None) and \
                             deriv.endswith(subject.record["modality"] + self.roi_params["suffix"] + ".nii.gz"):
