### 2020-12-02 ebf1c84
loader.py: change init target_filename
