### 2020-12-02 8f5b45d
First draft of HeMIS model implementation
