+++ ivadomed/models.py
@@ -615,6 +615,148 @@ class FiLMlayer(Module):
 
         return output, new_w_shared
 
+"""
+        Vanilla HeMIS structure:
+        BACK END
+        * 2x 2D-Convs (ReLU), stride 1
+        * (2,2) max pooling, stride 1
+        * Zero padding
+        ABSTRACTION
+        * Take mean/variation across modalities per feature map
+        FRONT END
+        * Concatenate mean and variation feature maps
+        * 2D-Conv (ReLU)
+        * 2D-Conv (Softmax, n_channels = n_classes)
+        """
+
+class HeMIS(Module):
+    """
+    An implementation of a vanilla HeMIS architecture (as shown in Havaei et al. 2016,
+    ArXiv link: https://arxiv.org/abs/1607.05194). This network deals with missing contrasts.
+    Optional arguments are present to include non-vanilla elements like dropout and batchnorm.
+
+    Param:
+        contrasts (list): list of all possible contrasts. ['T1', 'T2', 'T2S', 'F']
+
+    Args:
+        contrasts (list): List of contrasts.
+        out_channel (int): Number of output channels.
+        drop_rate (float): Probability of dropout. Set to None for no dropout.
+        bn_momentum (float): Batch normalization momentum. Set to None for no batchnorm.
+        **kwargs:
+
+    Attributes:
+        contrasts (list): List of contrasts.
+        branches (ModuleDict): Contains convolutional branch in the backend for each modality.
+    """
+
+    def __init__(self, contrasts, out_channel, drop_rate=None, bn_momentum=None):
+        super().__init__()
+        self.contrasts = contrasts
+
+        # Back end
+        self.branches = nn.ModuleDict(
+            [['branch_{}'.format(mod),
+              self.build_vanilla_backend(name=mod, 
+                                         bn_momentum=bn_momentum,
+                                         drop_rate=drop_rate)] for mod in self.contrasts])
+
+        # Front end
+        self.frontend = self.build_vanilla_frontend(out_channels=out_channel,
+                                                     bn_momentum=bn_momentum,
+                                                     drop_rate=drop_rate)
+
+    def build_vanilla_backend(self, name, bn_momentum, drop_rate):
+        layers = []
+
+        # TODO: Fix one-off shape error somewhere
+
+        # First layer
+        layers.append(['{}_conv0'.format(name), nn.Conv2d(in_channels=1, out_channels=48,
+                                                          kernel_size=5, padding=2)])
+        if bn_momentum:
+            layers.append(['{}_bn0'.format(name), nn.BatchNorm2d(48, momentum=bn_momentum)])
+        layers.append(['{}_relu0'.format(name), nn.ReLU(inplace=True)])
+        if drop_rate:
+            layers.append(['{}_drop0'.format(name), nn.Dropout2d(drop_rate)])
+
+        # Second layer
+        layers.append(['{}_conv1'.format(name), nn.Conv2d(in_channels=48, out_channels=48,
+                                                          kernel_size=5, padding=2)])
+        if bn_momentum:
+            layers.append(['{}_bn1'.format(name), nn.BatchNorm2d(48, momentum=bn_momentum)])
+        layers.append(['{}_relu1'.format(name), nn.ReLU(inplace=True)])
+        if drop_rate:
+            layers.append(['{}_drop1'.format(name), nn.Dropout2d(drop_rate)])
+
+        layers.append(['{}_maxpool'.format(name), nn.MaxPool2d(kernel_size=2, stride=1)])
+
+        branch = nn.Sequential(OrderedDict(layers))
+        return branch
+
+    def build_vanilla_frontend(self, out_channels, bn_momentum, drop_rate):
+        layers = []
+
[…]
