### 2020-12-02 74fe4da
Add missing attribute softmax
