### 2020-12-02 e3d18aa
Add parsing of sessions.tsv and scans.tsv
