### 2020-12-02 b3c9d34
Add filtering of dataframe based on derivatives availability
