### 2020-12-02 94a890b
Add filtering based on contrasts
