### 2020-12-02 a18dc84
Remove filtering based on data_type
