+++ ivadomed/loader/adaptative.py
@@ -402,6 +402,9 @@ class Bids_to_hdf5:
                 grp[key].attrs["data_shape"] = input_metadata['data_shape']
             if "bounding_box" in input_metadata.keys():
                 grp[key].attrs["bounding_box"] = input_metadata['bounding_box']
+            if "crop_params" in input_metadata.keys() and 'CenterCrop' in input_metadata['crop_params']:
+                # TODO save whole dict in crop_params
+                grp[key].attrs["crop_params"] = input_metadata['crop_params']['CenterCrop']
 
             # GT
             key = "gt/{}".format(contrast)
@@ -425,6 +428,9 @@ class Bids_to_hdf5:
                 grp[key].attrs["zooms"] = gt_metadata['zooms']
             if "data_shape" in gt_metadata.keys():
                 grp[key].attrs["data_shape"] = gt_metadata['data_shape']
+            if "crop_params" in gt_metadata.keys():
+                grp[key].create_dataset('crop_params', data=gt_metadata['crop_params'])
+                grp[key].attrs.create("crop_params", gt_metadata['crop_params'])
             if gt_metadata['bounding_box'] is not None:
                 grp[key].attrs["bounding_box"] = gt_metadata['bounding_box']
 
@@ -608,7 +614,8 @@ class HDF5Dataset:
                                                         .format(line['Subjects'], ct)].attrs.items()})
             metadata['slice_index'] = line["Slices"]
             metadata['missing_mod'] = missing_modalities
-            metadata['crop_params'] = {}
+            if 'crop_params' not in metadata:
+                metadata['crop_params'] = {}
             input_metadata.append(metadata)
 
         # GT
@@ -624,7 +631,8 @@ class HDF5Dataset:
             gt_metadata.append(imed_loader_utils.SampleMetadata({key: value for key, value in
                                                                  self.hdf5_file[line['gt/' +
                                                                  self.gt_contrast]].attrs.items()}))
-            gt_metadata[n_gt]['crop_params'] = {}
+            if 'crop_params' not in gt_metadata[n_gt]:
+                gt_metadata[n_gt]['crop_params'] = {}
 
         # ROI
         roi_img = []
@@ -641,7 +649,8 @@ class HDF5Dataset:
             roi_metadata.append(imed_loader_utils.SampleMetadata({key: value for key, value in
                                                                   self.hdf5_file[
                                                                       line['roi/' + self.roi_contrast]].attrs.items()}))
-            roi_metadata[0]['crop_params'] = {}
+            if 'crop_params' not in roi_metadata[0]:
+                roi_metadata[0]['crop_params'] = {}
 
         # Run transforms on ROI
         # ROI goes first because params of ROICrop are needed for the followings
