### 2020-12-03 d9631f5
segment_volume: use postprocessing in config file
