### 2020-12-03 93042e3
apply safety factor to bounding box prior
