### 2020-12-03 e6367b0
loader.py: typo in comment
