### 2020-12-03 c2da7ba
loader.py: gthandle list for all
