### 2020-12-03 9da17fc
loader.py: get pair shapes
