### 2020-12-03 9962cc1
loader.py: as_closest_canonical
