### 2020-12-04 4b406ff
loader.py: rm old comment
