### 2020-12-04 a7aa7b5
loader.py: enumerate gt_rater
