### 2020-12-04 d90a6ea
loader.py: idx_tissue idx_rater
