+++ ivadomed/loader/loader.py
@@ -271,7 +271,7 @@ class SegmentationPair(object):
             dict: Input and gt metadata.
         """
         gt_meta_dict = []
-        for gt in self.gt_handle:
+        for idx_tissue, gt in enumerate(self.gt_handle):
             if gt is not None:
                 if not isinstance(gt, list):  # this tissue has annotation from only one rater
                     gt_meta_dict.append(imed_loader_utils.SampleMetadata({
@@ -286,11 +286,11 @@ class SegmentationPair(object):
                     gt_meta_dict.append([imed_loader_utils.SampleMetadata({
                         "zooms": imed_loader_utils.orient_shapes_hwd(gt_rater.header.get_zooms(), self.slice_axis),
                         "data_shape": imed_loader_utils.orient_shapes_hwd(gt_rater.header.get_data_shape(), self.slice_axis),
-                        "gt_filenames": self.metadata[0]["gt_filenames"][idx],
+                        "gt_filenames": self.metadata[0]["gt_filenames"][idx_tissue][idx_rater],
                         "bounding_box": self.metadata[0]["bounding_box"] if 'bounding_box' in self.metadata[0] else None,
                         "data_type": 'gt',
                         "crop_params": {}
-                    }) for idx, gt_rater in enumerate(gt)])
+                    }) for idx_rater, gt_rater in enumerate(gt)])
 
             else:
                 # Temporarily append null metadata to null gt
