### 2020-12-04 a710ba8
loader.py: gt_obj
