+++ ivadomed/loader/utils.py
@@ -445,7 +445,11 @@ def update_metadata(metadata_src_lst, metadata_dest_lst):
         list: updated metadata list.
     """
     if metadata_src_lst and metadata_dest_lst:
-        metadata_dest_lst[0]._update(metadata_src_lst[0], TRANSFORM_PARAMS)
+        if not isinstance(metadata_dest_lst[0], list):
+            metadata_dest_lst[0]._update(metadata_src_lst[0], TRANSFORM_PARAMS)
+        else:
+            for idx, _ in enumerate(metadata_dest_lst[0]):
+                metadata_dest_lst[0][idx]._update(metadata_src_lst[0][idx], TRANSFORM_PARAMS)
     return metadata_dest_lst
 
 
