### 2020-12-04 1e235e6
loader/utils: metadata shape
