### 2020-12-04 62c3422
loader.py: randomly pick rater
