### 2020-12-04 c0d32ef
transforms.py: rm multi_capable unneeded
