### 2020-12-04 a1ffc5e
main.py: multiple rater: only for training
