### 2020-12-04 736f1ee
don't discard subjects that misses some GT for multichannel
