### 2020-12-04 12a5b50
Update force_index condition for microscopy files
