### 2020-12-04 cc3e822
minor fix in loader
