### 2020-12-04 dd1a5ec
Use bids dataset in adaptive (reduce code duplication)
