+++ ivadomed/loader/adaptative.py
@@ -175,26 +175,17 @@ class Bids_to_hdf5:
     """Converts a BIDS dataset to a HDF5 file.
 
     Args:
-        root_dir (str): Path to the BIDS dataset.
-        subject_lst (list): Subject names list.
-        target_suffix (list): List of suffixes for target masks.
-        roi_params (dict): Dictionary containing parameters related to ROI image processing.
-        contrast_lst (list): List of the contrasts.
+        dataset (BidsDataset): Bids dataset.
         hdf5_name (str): Path and name of the hdf5 file.
-        contrast_balance (dict): Dictionary controlling image contrasts balance.
         slice_axis (int): Indicates the axis used to extract slices: "axial": 2, "sagittal": 0, "coronal": 1.
         metadata_choice (str): Choice between "mri_params", "contrasts", None or False, related to FiLM.
         slice_filter_fn (SliceFilter): Class that filters slices according to their content.
-        transform (Compose): Transformations.
-        object_detection_params (dict): Object detection parameters.
 
     Attributes:
-        bids_ds (BIDS): BIDS dataset.
         dt (dtype): hdf5 special dtype.
         hdf5_file (hdf5): hdf5 file containing dataset information.
         filename_pairs (list): A list of tuples in the format (input filename list containing all modalities,ground \
             truth filename, ROI filename, metadata).
-        metadata (dict): Dictionary containing metadata of input and gt.
         prepro_transforms (Compose): Transforms to be applied before training.
         transform (Compose): Transforms to be applied during training.
         has_bounding_box (bool): True if all metadata contains bounding box coordinates, else False.
@@ -202,95 +193,29 @@ class Bids_to_hdf5:
         slice_filter_fn (SliceFilter): Object that filters slices according to their content.
     """
 
-    def __init__(self, root_dir, subject_lst, target_suffix, contrast_lst, hdf5_name, contrast_balance=None,
-                 slice_axis=2, metadata_choice=False, slice_filter_fn=None, roi_params=None, transform=None,
-                 object_detection_params=None, soft_gt=False):
+    def __init__(self,dataset, hdf5_name, slice_axis=2, metadata_choice=False, slice_filter_fn=None, soft_gt=False):
         print("Starting conversion")
-        # Getting all patients id
-        self.bids_ds = bids.BIDS(root_dir)
-        bids_subjects = [s for s in self.bids_ds.get_subjects() if s.record["subject_id"] in subject_lst]
         self.soft_gt = soft_gt
         self.dt = h5py.special_dtype(vlen=str)
         # opening an hdf5 file with write access and writing metadata
         self.hdf5_file = h5py.File(hdf5_name, "w")
+        self.has_bounding_box = True
 
         list_patients = []
 
-        self.filename_pairs = []
+        self.filename_pairs = dataset.filename_pairs
+        self.prepro_transforms = dataset.prepro_transforms
+        self.transform = dataset.transform
 
-        if metadata_choice == 'mri_params':
-            self.metadata = {"FlipAngle": [], "RepetitionTime": [],
-                             "EchoTime": [], "Manufacturer": []}
-
-        self.prepro_transforms, self.transform = transform
-        # Create a list with the filenames for all contrasts and subjects
-        subjects_tot = []
-        for subject in bids_subjects:
-            subjects_tot.append(str(subject.record["absolute_path"]))
-
-        # Create a dictionary with the number of subjects for each contrast of contrast_balance
-        tot = {contrast: len([s for s in bids_subjects if s.record["modality"] == contrast])
-               for contrast in contrast_balance.keys()}
-
-        # Create a counter that helps to balance the contrasts
[…]
