### 2020-12-04 f8ee80d
create one hdf5 file per dataset type
