### 2020-12-04 faf6540
Add warning when no derivatives exist for a subject file
