### 2020-12-04 b0c798b
update HeMIS and adaptive tests
