+++ testing/test_HeMIS.py
@@ -59,7 +59,7 @@ def test_HeMIS(p=0.0001):
             "out_channel": 1,
             "missing_probability": 0.00001,
             "missing_probability_growth": 0.9,
-            "contrasts": ["T1w", "T2w"],
+            "contrasts": ["T1w", "T2w", "T2star"],
             "ram": False,
             "hdf5_path": 'testing_data/mytestfile.hdf5',
             "csv_path": 'testing_data/hdf5.csv',
@@ -78,6 +78,7 @@ def test_HeMIS(p=0.0001):
                                           slice_axis=2,
                                           transform=transform_lst,
                                           metadata_choice=False,
+                                          multichannel=True,
                                           dim=2,
                                           slice_filter_fn=imed_loader_utils.SliceFilter(filter_empty_input=True,
                                                                                  filter_empty_mask=True),
