### 2020-12-04 7c3e8a2
Add warning when a target_suffix is missing for a subject file
