### 2020-12-05 10eacf8
Update models.py
The 2D attention module is added to the code.
the headlines of the changes are elaborated in the following:
1. The UnetGridGatingSignal2 is added due to dealing with 2D input.
2. the _GridAttentionBlockND is rewrite based on the original code posted on GitHub (https://github.com/ozan-oktay/Attention-Gated-Networks) and now is compatible with 2D and 3D input.
3. In the Decoder class, the attention module is added.
4. some minor changes in the decoder and encoder class have been made due to make the code more modular (e.g., with depth changes, etc.)
