+++ ivadomed/models.py
@@ -293,7 +293,6 @@ class UpConv(Module):
         x = self.downconv(x)
         return x
 
-
 class Encoder(Module):
     """Encoding part of the U-Net model.
     It returns the features map for the skip connections
@@ -325,14 +324,13 @@ class Encoder(Module):
         self.down_path.append(max_pool(2))
 
         for i in range(depth - 1):
-            self.down_path.append(DownConv(n_filters, n_filters * 2, drop_rate, bn_momentum, is_2d))
-            self.down_path.append(FiLMlayer(n_metadata, n_filters * 2) if film_layers and film_layers[i + 1] else None)
+            self.down_path.append(DownConv(n_filters*2**i, n_filters*2**(i+1), drop_rate, bn_momentum, is_2d))
+            self.down_path.append(FiLMlayer(n_metadata, n_filters * 2**(i+1)) if film_layers and film_layers[i + 1] else None)
             self.down_path.append(max_pool(2))
-            n_filters = n_filters * 2
 
         # Bottom
-        self.conv_bottom = DownConv(n_filters, n_filters, drop_rate, bn_momentum, is_2d)
-        self.film_bottom = FiLMlayer(n_metadata, n_filters) if film_layers and film_layers[self.depth] else None
+        self.conv_bottom = DownConv(n_filters*2**(depth-1), n_filters*2**depth, drop_rate, bn_momentum, is_2d)
+        self.film_bottom = FiLMlayer(n_metadata, n_filters*2**depth) if film_layers and film_layers[self.depth] else None 
 
     def forward(self, x, context=None):
         features = []
@@ -357,10 +355,8 @@ class Encoder(Module):
         if self.film_bottom:
             x, w_film = self.film_bottom(x, context, None if 'w_film' not in locals() else w_film)
         features.append(x)
-
         return features, None if 'w_film' not in locals() else w_film
-
-
+    
 class Decoder(Module):
     """Decoding part of the U-Net model.
 
@@ -373,6 +369,7 @@ class Decoder(Module):
         film_layers (list): List of 0 or 1 indicating on which layer FiLM is applied.
         hemis (bool): Boolean indicating if HeMIS is on or not.
         final_activation (str): Choice of final activation between "sigmoid", "relu" and "softmax"
+        attention (bool): Boolian indicating if attention module is on or not.
 
     Attributes:
         depth (int): Number of down convolutions minus bottom down convolution.
@@ -383,14 +380,19 @@ class Decoder(Module):
         softmax (Softmax): Softmax layer that can be applied as last layer.
     """
 
+
     def __init__(self, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
-                 n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True, n_filters=64):
+                 n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True,
+                 n_filters=64, attention= True):
         super(Decoder, self).__init__()
         self.depth = depth
         self.out_channel = out_channel
+        self.base_n_filter = n_filters
+        self.attention = attention
         self.final_activation = final_activation
         # Up-Sampling path
         self.up_path = nn.ModuleList()
+        self.att_path = nn.ModuleList()
         if hemis:
             in_channel = n_filters * 2 ** self.depth
             self.up_path.append(UpConv(in_channel * 2, n_filters * 2 ** (self.depth - 1), drop_rate, bn_momentum,
@@ -403,7 +405,8 @@ class Decoder(Module):
         else:
             in_channel = n_filters * 2 ** self.depth
 
[…]
