### 2020-12-06 19bf736
Update test_model.py
The unit test for the evaluation of Encoder, Decoder, and 2D attention module added.
