### 2020-12-06 67614a4
Update test_model.py
