+++ testing/test_model.py
@@ -52,7 +52,7 @@ def test_film_generator():
     assert (type(inf[0]) == torch.Tensor)
     assert (type(inf[1]) == torch.nn.parameter.Parameter)
 
-def test_modified_Encoder():
+def test_modified_encoder():
     # verifying changes in the Encoder class
     a = [[[[0 for i in range(322)]for i in range(322)]]]
     inp = torch.tensor(a).float()
@@ -61,18 +61,18 @@ def test_modified_Encoder():
     inf = model(inp)
     assert(type(inf) == torch.Tensor)
     
-def test_modified_Decoder():
+def test_modified_decoder():
     # verifying if changes such as 2D attention and other modification works 
     a = [torch.zeros((1, 64, 322, 322)), torch.zeros((1, 128, 161, 161)),
          torch.zeros((1, 256, 80, 80)), torch.zeros((1, 512, 40, 40))] # by changing the depth, the length of the input feature list should be changed
-    inp = a # should not be transformed to torch tensor because "a" is the appended feature list
+    inp = a # should not be transformed to torch tensor because "a" is the appended feature list from skip connections
     model = imed_model.Decoder(out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
                  n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True,
-                 n_filters=64, attention= True)
+                 n_filters=64, attention=True)
     inf = model(inp)
     assert(type(inf) == torch.Tensor)
     
-def test_Unet_att2D():
+def test_unet_att2d():
     
     # verifying if the 2D attention works 
     a = [[[[0 for i in range(322)]for i in range(322)]]]
