### 2020-12-06 9ba5fa8
Update test_model.py
