### 2020-12-06 1d1881b
Update test_model.py
For the encoder class unit test, the output of the model should be tuple, not torch.Tensor.
Because the output of the encoder section is a tuple consist of a list from the skip connection (features and bottom layer) and also a None from the last line of code 
(features, None if 'w_film' not in locals() else w_film)
