### 2020-12-07 07a54ea
loader.py: random expert for 3D
