+++ ivadomed/main.py
@@ -178,9 +178,11 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
 
     # Check if multiple raters
     if any([class_suffix for class_suffix in loader_params["target_suffix"]]):
-        print("\nAnnotations from multiple raters will be used during model training, one annotation from one rater randomly selected at each iteration.\n")
+        print(
+            "\nAnnotations from multiple raters will be used during model training, one annotation from one rater randomly selected at each iteration.\n")
         if command != "train":
-            print("\nERROR: Please provide only one annotation per class in 'target_suffix' when not training a model.\n")
+            print(
+                "\nERROR: Please provide only one annotation per class in 'target_suffix' when not training a model.\n")
             exit()
 
     if command == 'train':
@@ -245,7 +247,8 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
         if command != 'train':  # If command == train, then ds_valid already load
             # Get Validation dataset
             ds_valid = imed_loader.load_dataset(**{**loader_params,
-                                                   **{'data_list': valid_lst, 'transforms_params': transform_valid_params,
+                                                   **{'data_list': valid_lst,
+                                                      'transforms_params': transform_valid_params,
                                                       'dataset_type': 'validation'}}, device=device,
                                                 cuda_available=cuda_available)
         # Get Training dataset with no Data Augmentation
@@ -287,7 +290,7 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
                                                                   'transforms_params': transformation_dict,
                                                                   'dataset_type': 'testing',
                                                                   'requires_undo': True}}, device=device,
-                                                                  cuda_available=cuda_available)
+                                           cuda_available=cuda_available)
 
         metric_fns = imed_metrics.get_metric_fns(ds_test.task)
 
