### 2020-12-07 4ffb898
loader.py: deal w unlabelled data
