### 2020-12-07 7a5e00e
loader.py: old init for target_filena,es
