### 2020-12-07 414a15c
Implementation of max pooling (#551)
* implementation of max pooling

* Add documentation in RTD
