### 2020-12-07 #2868 @justusschock
* __Installation:__ With Python 3.9 it does not work, although the documentation states, it should just be Python>=3.6
```bash
pip install ivadomed
Collecting ivadomed
  Downloading ivadomed-2.6.0-py3-none-any.whl (139 kB)
     |████████████████████████████████| 139 kB 8.9 MB/s
  Downloading ivadomed-2.5.0-py3-none-any.whl (137 kB)
     |████████████████████████████████| 137 kB 17.9 MB/s
  Downloading ivadomed-2.4.0-py3-none-any.whl (132 kB)
     |████████████████████████████████| 132 kB 23.7 MB/s
  Downloading ivadomed-2.3.1-py3-none-any.whl (129 kB)
     |████████████████████████████████| 129 kB 24.1 MB/s
  Downloading ivadomed-2.2.1-py3-none-any.whl (127 kB)
     |████████████████████████████████| 127 kB 28.2 MB/s
  Downloading ivadomed-2.2.0-py3-none-any.whl (127 kB)
     |████████████████████████████████| 127 kB 23.9 MB/s
  Downloading ivadomed-2.1.0-py3-none-any.whl (110 kB)
     |████████████████████████████████| 110 kB 30.8 MB/s
  Downloading ivadomed-2.0.2-py3-none-any.whl (108 kB)
     |████████████████████████████████| 108 kB 29.9 MB/s
  Downloading ivadomed-2.0.1-py3-none-any.whl (108 kB)
     |████████████████████████████████| 108 kB 30.9 MB/s
  Downloading ivadomed-1.2.1-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 1.8 MB/s
  Downloading ivadomed-1.2-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 5.0 MB/s
  Downloading ivadomed-1.1-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 4.6 MB/s
  Downloading ivadomed-1.0-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 2.4 MB/s
  Downloading ivadomed-0.14-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 4.8 MB/s
  Downloading ivadomed-0.13-py3-none-any.whl (68 kB)
     |████████████████████████████████| 68 kB 457 kB/s
  Downloading ivadomed-0.12-py3-none-any.whl (66 kB)
     |████████████████████████████████| 66 kB 3.5 MB/s
  Downloading ivadomed-0.11-py3-none-any.whl (66 kB)
     |████████████████████████████████| 66 kB 3.3 MB/s
ERROR: Cannot install ivadomed==0.11, ivadomed==0.12, ivadomed==0.13, ivadomed==0.14, ivadomed==1.0, ivadomed==1.1, ivadomed==1.2, ivadomed==1.2.1, ivadomed==2.0.1, ivadomed==2.0.2, ivadomed==2.1.0, ivadomed==2.2.0, ivadomed==2.2.1, ivadomed==2.3.1, ivadomed==2.4.0, ivadomed==2.5.0 and ivadomed==2.6.0 because these package versions have conflicting dependencies.

The conflict is caused by:
    ivadomed 2.6.0 depends on onnxruntime==1.4.0
    ivadomed 2.5.0 depends on onnxruntime==1.4.0
    ivadomed 2.4.0 depends on onnxruntime
    ivadomed 2.3.1 depends on torchvision>=0.6.0
    ivadomed 2.2.1 depends on torchvision>=0.6.0
    ivadomed 2.2.0 depends on torchvision>=0.6.0
    ivadomed 2.1.0 depends on torchvision>=0.6.0
    ivadomed 2.0.2 depends on torchvision>=0.6.0
    ivadomed 2.0.1 depends on torchvision>=0.6.0
    ivadomed 1.2.1 depends on torchvision>=0.6.0
    ivadomed 1.2 depends on torchvision>=0.6.0
    ivadomed 1.1 depends on torchvision>=0.6.0
    ivadomed 1.0 depends on torchvision>=0.6.0
    ivadomed 0.14 depends on torchvision>=0.6.0
    ivadomed 0.13 depends on torchvision>=0.6.0
    ivadomed 0.12 depends on torchvision>=0.6.0
    ivadomed 0.11 depends on torchvision>=0.6.0

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
```
With Python 3.8.5 it works fine! Maybe having the installation instructions also in the readme and not only in the linked documentation would be nice.

* __State of Need:__
While it is clearly stated, why the software is necessary and how it compares to other available software, the table at https://ivadomed.org/en/latest/purpose.html#comparison-with-other-projects is not nice to look at. Maybe divide it into several tables with major parts.

* __Automated tests__: While tests are present and coverage is reported correctly, it would be nice having a badge pointing to the actual outcome of GitHub actions.

* __Community Guidelines__: While they are available at https://ivadomed.org/en/latest/contributing.html and in the repository wiki, it is not the usual place for GitHub repositories. SO it would be nice to also have a file like `CODE_OF_CONDUCT.md` (which must only point to the available resources) as well as issue and Pull-Request templates pointing to the relevant parts of the wiki.
