### 2020-12-07 6b8eeb1
Correction to filtering when no extensions are provided
