### 2020-12-07 baf1fd5
PEP8 follow test_model.py
* applying PEP8 with Ctrl+Alt+L using PyCharm
