+++ testing/test_model.py
@@ -2,6 +2,7 @@ import ivadomed.models as imed_model
 import torch
 import torchvision
 
+
 # testing countception model
 def test_countception():
     a = [[[[0 for i in range(10)] for i in range(10)]]]
@@ -17,7 +18,7 @@ def test_model_3d_att():
     inp = torch.tensor(a).float()
     model = imed_model.Modified3DUNet(in_channel=1, out_channel=1, attention=True)
     inf = model(inp)
-    assert(type(inf) == torch.Tensor)
+    assert (type(inf) == torch.Tensor)
 
 
 def test_resnet():
@@ -52,35 +53,36 @@ def test_film_generator():
     assert (type(inf[0]) == torch.Tensor)
     assert (type(inf[1]) == torch.nn.parameter.Parameter)
 
+
 def test_modified_encoder():
     # verifying changes in the Encoder class
-    a = [[[[0 for i in range(322)]for i in range(322)]]]
+    a = [[[[0 for i in range(322)] for i in range(322)]]]
     inp = torch.tensor(a).float()
     model = imed_model.Encoder(in_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1, n_metadata=None, film_layers=None,
-                 is_2d=True, n_filters=64)
+                               is_2d=True, n_filters=64)
     inf = model(inp)
-    assert(type(inf) == tuple)
-    
+    assert (type(inf) == tuple)
+
+
 def test_modified_decoder():
-    # verifying if changes such as 2D attention and other modification works 
+    # verifying if changes such as 2D attention and other modification works
     a = [torch.zeros((1, 64, 322, 322)), torch.zeros((1, 128, 161, 161)),
-         torch.zeros((1, 256, 80, 80)), torch.zeros((1, 512, 40, 40))] # by changing the depth, the length of the input feature list should be changed.
-    inp = a # should not be transformed to torch tensor because "a" is the appended feature list from the skip connections.
+         torch.zeros((1, 256, 80, 80)), torch.zeros(
+            (1, 512, 40, 40))]  # by changing the depth, the length of the input feature list should be changed.
+    inp = a  # should not be transformed to torch tensor because "a" is the appended feature list from the skip connections.
     model = imed_model.Decoder(out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
-                 n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True,
-                 n_filters=64, attention=True)
+                               n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True,
+                               n_filters=64, attention=True)
     inf = model(inp)
-    assert(type(inf) == torch.Tensor)
-    
+    assert (type(inf) == torch.Tensor)
+
+
 def test_unet_att2d():
-    
-    # verifying if the 2D attention works 
-    a = [[[[0 for i in range(322)]for i in range(322)]]]
+    # verifying if the 2D attention works
+    a = [[[[0 for i in range(322)] for i in range(322)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.Unet(in_channel=1, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1, final_activation='sigmoid',
-                 is_2d=True, n_filters=64)
+    model = imed_model.Unet(in_channel=1, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
+                            final_activation='sigmoid',
+                            is_2d=True, n_filters=64)
     inf = model(inp)
-    assert(type(inf) == torch.Tensor)
-
-
-
+    assert (type(inf) == torch.Tensor)
