### 2020-12-07 bd307fe
loader.py: rm draft lines
