### 2020-12-07 84c4e01
Update test_model.py
* related testing units changes for 2d attention module is done [on this branch](https://github.com/AshkanTaghipour/ivadomed/blob/patch-1/testing/test_model.py)
* previously test units deleted
