### 2020-12-07 0124fac
Update models.py
* follow PEP8
