+++ ivadomed/models.py
@@ -9,12 +9,11 @@ from torch.nn import init
 import torchvision.models
 
 
-
-#Modified from torchvision.models.resnet.Resnet
+# Modified from torchvision.models.resnet.Resnet
 class ResNet(nn.Module):
     """ResNet model based on
      `"Deep Residual Learning for Image Recognition" <https://arxiv.org/abs/1512.03385.pdf>`_
-    
+
     Args:
         block (nn.Module): Basic block of the network (such as conv + bn + non-nonlinearity)
         layers (int list): Number of blocks to stack (network depth) after each downsampling step.
@@ -27,6 +26,7 @@ class ResNet(nn.Module):
         norm_layer (layer): Custom normalization layer, if not provided BatchNorm2d is used
 
     """
+
     def __init__(self, block, layers, num_classes=2, zero_init_residual=False,
                  groups=1, width_per_group=64, replace_stride_with_dilation=None,
                  norm_layer=None):
@@ -177,7 +177,7 @@ class DenseNet(nn.Module):
             num_features = num_features + num_layers * growth_rate
             if i != len(block_config) - 1:
                 trans = torchvision.models.densenet._Transition(num_input_features=num_features,
-                                    num_output_features=num_features // 2)
+                                                                num_output_features=num_features // 2)
                 self.features.add_module('transition%d' % (i + 1), trans)
                 num_features = num_features // 2
 
@@ -293,6 +293,7 @@ class UpConv(Module):
         x = self.downconv(x)
         return x
 
+
 class Encoder(Module):
     """Encoding part of the U-Net model.
     It returns the features map for the skip connections
@@ -324,13 +325,15 @@ class Encoder(Module):
         self.down_path.append(max_pool(2))
 
         for i in range(depth - 1):
-            self.down_path.append(DownConv(n_filters*2**i, n_filters*2**(i+1), drop_rate, bn_momentum, is_2d))
-            self.down_path.append(FiLMlayer(n_metadata, n_filters * 2**(i+1)) if film_layers and film_layers[i + 1] else None)
+            self.down_path.append(DownConv(n_filters * 2 ** i, n_filters * 2 ** (i + 1), drop_rate, bn_momentum, is_2d))
+            self.down_path.append(
+                FiLMlayer(n_metadata, n_filters * 2 ** (i + 1)) if film_layers and film_layers[i + 1] else None)
             self.down_path.append(max_pool(2))
 
         # Bottom
-        self.conv_bottom = DownConv(n_filters*2**(depth-1), n_filters*2**depth, drop_rate, bn_momentum, is_2d)
-        self.film_bottom = FiLMlayer(n_metadata, n_filters*2**depth) if film_layers and film_layers[self.depth] else None 
+        self.conv_bottom = DownConv(n_filters * 2 ** (depth - 1), n_filters * 2 ** depth, drop_rate, bn_momentum, is_2d)
+        self.film_bottom = FiLMlayer(n_metadata, n_filters * 2 ** depth) if film_layers and film_layers[
+            self.depth] else None
 
     def forward(self, x, context=None):
         features = []
@@ -356,7 +359,8 @@ class Encoder(Module):
             x, w_film = self.film_bottom(x, context, None if 'w_film' not in locals() else w_film)
         features.append(x)
         return features, None if 'w_film' not in locals() else w_film
-    
+
+
 class Decoder(Module):
     """Decoding part of the U-Net model.
 
@@ -380,10 +384,9 @@ class Decoder(Module):
         softmax (Softmax): Softmax layer that can be applied as last layer.
     """
 
-
     def __init__(self, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
                  n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True,
[…]
