### 2020-12-08 ad6473e
Handling python version for install (#562)
* update with information about python version as well as torchvision

* update readmed with Python 3.9

* change option
