### 2020-12-08 e67f0fc
Updating comparison table after review (#560)
* trying a new table division

* update after few mistakes
