### 2020-12-08 7d3eef7
Add create_bids_dataframe function
* requirements: Add pybids>=0.12.4

* config: Add custom config_bids.json file

* config: Add config_new_loader.json

* loader/utils.py: Add create_bids_dataframe function

* dev: Add df_new_loader.py script to test dataframe function

* Fix typo

* config: use relative paths

* Define path with os.join

* Remove unneccesary print

* bids_config default to None if not present in config file

* Change file filtering to work with relative paths and avoid df concat

* Read participants.tsv only if present

* Correction PEP8

* Minor corrections script dev

* Con
[…]
