### 2020-12-08 f471253
Update models.py
* PEP8 follow
* replacing F.upsampling by F.interpolate
* replacing F.sigmoid by torch.sigmoid
* description about the forward path of attention gate module
