### 2020-12-08 96ed0c3
Update models type error avoid
* replace NotImplementedby NotImplementedError to avoid type errors
