### 2020-12-08 6ab643c
Update model AttributeError
* try to removing the  AttributeError by correct indenting the _concatenation function
