### 2020-12-08 373feb5
modify multiple rater condition
