+++ README.md
@@ -3,7 +3,7 @@
 
 [![Coverage Status](https://coveralls.io/repos/github/ivadomed/ivadomed/badge.svg?branch=master)](https://coveralls.io/github/ivadomed/ivadomed?branch=master)
 ![](https://github.com/neuropoly/ivadomed/workflows/Python%20package/badge.svg)
-[![test status](https://github.com/ivadomed/ivadomed/workflows/.github/workflows/run_test.yml/badge.svg)](https://github.com/ivadomed/ivadomed/actions)
+[![test status](https://github.com/ivadomed/ivadomed/workflows/.github/workflows/run_tests.yml/badge.svg)](https://github.com/ivadomed/ivadomed/actions)
 [![Documentation Status](https://readthedocs.org/projects/ivado-medical-imaging/badge/?version=stable)](https://ivadomed.org/en/stable/?badge=stable)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE.md)
 
