### 2020-12-08 09dfbb5
contributing.rst: Change to symlink to root-level file
