### 2020-12-08 d583600
CODE_OF_CONDUCT.md: Adopt Contributor Covenant
Source: https://www.contributor-covenant.org/
