### 2020-12-08 4501109
Update test_HeMIS.py
* the depths' number changes from 2 to 3
