### 2020-12-08 7631f81
feature_request.md: Adapt SCT feature request template
