+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,25 @@
+---
+name: Feature request
+about: Want improvements or new features in ivadomed?
+
+---
+
+<!-- Hi, and thank you for requesting a feature! Please take the time to first consider NeuroPoly's guidelines on issues (titles, descriptions, issue labels) before continuing:
+https://www.neuro.polymtl.ca/software/contributing#opening_an_issue
+-->
+
+### Motivation for the feature
+<!-- Is your feature request related to a problem? Please provide a clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+-->
+
+### Description of the feature
+<!-- A clear and concise description of the feature you would like added, and any implementation details you've thought of.
+-->
+
+### Alternatives
+<!-- A clear and concise description of any alternative solutions or features you've considered.
+-->
+
+### Additional context
+<!-- Add any other context (references, screenshots...) about the feature request here.
+-->
