### 2020-12-08 35b23cc
use tranpose csv instead of splitting the table
