### 2020-12-08 afc82c2
remove unneeded badge
