### 2020-12-09 7c45854
CODE_OF_CONDUCT.md: Add contact information
