### 2020-12-09 94c2afe
test now cleans after himself(download_data)
