### 2020-12-09 972823c
test now cleans after himself(segment and onnx convert)
