### 2020-12-09 22c81bd
test now cleans after himself(visualize and extract_small_ds)
