### 2020-12-09 db9e165
test now cleans after himself(hyperparamopt)
