### 2020-12-09 069fac4
error in automate training path logdir
