+++ testing/test_script.py
@@ -219,6 +219,10 @@ def test_automate_training_train():
     command = "ivadomed_automate_training -c testing_data/model_config_auto.json " \
               "-p testing_data/hyperparameter_opt.json -n 1 --fixed-split"
     subprocess.check_output(command, shell=True)
+    assert os.path.isdir("testing_script-batch_size=2")
+    assert os.path.isdir("testing_script-batch_size=4")
+    shutil.rmtree("testing_script-batch_size=2")
+    shutil.rmtree("testing_script-batch_size=4")
 
 
 def test_training_curve_multiple():
@@ -230,10 +234,7 @@ def test_automate_training_test():
     command = "ivadomed_automate_training -c testing_data/model_config_auto.json " \
               "-p testing_data/hyperparameter_opt.json -n 1 --run-test --all-combin -t 0.1"
     subprocess.check_output(command, shell=True)
-    assert os.path.isdir("testing_script-batch_size\=2")
-    assert os.path.isdir("testing_script-batch_size\=4")
-    shutil.rmtree("testing_script-batch_size\=2")
-    shutil.rmtree("testing_script-batch_size\=4")
+
 
 
 def test_create_json_3d_unet_test():
