### 2020-12-09 c12a0ac
move cleaning at the end of the file
