### 2020-12-09 db6f622
add cleaning at th end of the test
