### 2020-12-09 7e2c1b6
move cleaning at the end of the file (modification to be consistent)
