### 2020-12-09 3a36a56
Adapt remove small for multiclass
