### 2020-12-09 cbb6855
remove_small: convert int thr to list
