### 2020-12-09 9198b97
move cleaning at the end of the file (modification to be consistent)
