### 2020-12-09 6729cfd
segment_volume: adapt for multiclass remove small
