### 2020-12-09 7d6af93
delete postprocessing if set to 0
