### 2020-12-10 147d8a7
models Hemis Update
* the input number of Hemis conditioned halved, due to doubling the number of feature maps in the bottom layer of encoder class
