### 2020-12-10 1eb2b18
Added README for getting started with the Sphinx docs
