### 2020-12-10 10ddbdf
Hemis Condition fixed
* the number of input channel for Hemis condition set as the same as the previous version (in_channel*final output class)
