### 2020-12-10 1d794eb
Add help section to Sphinx docs
