+++ docs/source/help.rst
@@ -1,0 +1,6 @@
+Help
+====
+
+If you need help using ``ivadomed``, please don't hesitate to
+`post an issue on GitHub <https://github.com/ivadomed/ivadomed/issues>`__
+and we will be happy to assist you.
