### 2020-12-10 8945272
Added context about BIDS
