+++ docs/source/purpose.rst
@@ -36,10 +36,9 @@ inconsistencies, please let us know by `opening an issue <https://github.com/iva
 .. csv-table::
    :file: comparison_other_projects_table.csv
 
-(1): `Brain Imaging Data Structure <https://bids.neuroimaging.io/>`_ 
+(1): `"BIDS" stands for the Brain Imaging Data Structure <https://bids.neuroimaging.io/>`_, which is a convention initiated by the neuroimaging community to organize datasets (filenames, metadata, etc.). This facilitates the sharing of datasets and minimizes the burden of organizing datasets for training.
 
 (2): Class: Classification | Seg: Segmentation | Detect: Detection | Gen: Generation | Clust: Clustering | Reg: Registration
-
 
 
 
