### 2020-12-10 31dfda4
Fixed display of hyperlink
