### 2020-12-10 #2868 @charleygros
Many thanks for taking the time to review our work and providing valuable feedback @justusschock !

> * **Installation:** With Python 3.9 it does not work, although the documentation states, it should just be Python>=3.6

Thank you for raising this point! We modified our documentation accordingly, see [PR #562](https://github.com/ivadomed/ivadomed/pull/562), stating that ivadomed works for Python >= 3.6 and < 3.9.

> With Python 3.8.5 it works fine! Maybe having the installation instructions also in the readme and not only in the linked documentation would be nice.

We agree, thanks for suggesting this. We added the installation instructions in the README as well, see [PR #561](https://github.com/ivadomed/ivadomed/pull/561).

>   While it is clearly stated, why the software is necessary and how it compares to other available software, the table at https://ivadomed.org/en/latest/purpose.html#comparison-with-other-projects is not nice to look at. Maybe divide it into several tables with major parts.

We modified the layout of the table by transposing it and making them more compact, see [PR #567](https://github.com/ivadomed/ivadomed/pull/567).

> While tests are present and coverage is reported correctly, it would be nice having a badge pointing to the actual outcome of GitHub actions.

We added badges pointing to the Github action panels in [PR #566](https://github.com/ivadomed/ivadomed/pull/566).

> While they are available at https://ivadomed.org/en/latest/contributing.html and in the repository wiki, it is not the usual place for GitHub repositories. SO it would be nice to also have a file like `CODE_OF_CONDUCT.md` (which must only point to the available resources) as well as issue and Pull-Request templates pointing to the relevant parts of the wiki.

We agree with you @justusschock, thank you for raising this point. In the [PR #565](https://github.com/ivadomed/ivadomed/pull/565), we did the following changes:
- Add 3 templates (1 for PRs and 2 for issues)
- Add a CODE_OF_CONDUCT.md that uses the Contributor Covenant
- Move the contributing.rst to the project root to make it more visible, and add a symlink within the docs, so that it can appear in both places at once.
