### 2020-12-11 #2868 @lbugnon
Hi @charleygros, just a minor thing,  `virtualenv venv-ivadomed --python=python3.6` gives an error (because i have installed python 3.7, I think It could be omited or stated to look into the OS.
