### 2020-12-11 a2ebd46
Merge 2 BIDS datasets. Subjects' files are copied to a new directory. participants.tsv and participants.json are merged into one
