### 2020-12-11 #2868 @lbugnon
Usage (as a demo) is not clear, for instance `ivadomed ivadomed/config/config.json` doesn't work. Adding -c option lead to error RuntimeError: The dataset_description.json file was not found.
