### 2020-12-11 #2868 @lbugnon
@charleygros referneces are missing in the paper md
