### 2020-12-11 c439890
Apply bounding box safety factor in segment volume (#549)
* segment_volume: use postprocessing in config file

* apply safety factor to bounding box prior

* overwrite postprocessing step by user values

* fix typo

* delete postprocessing if set to 0

Co-authored-by: lrouhier <lucas.rouhier@telecom-paristech.fr>
