### 2020-12-11 3004dfe
Remove small for multiclass (#570)
* Adapt remove small for multiclass

* remove_small: convert int thr to list

* add documentation

* segment_volume: adapt for multiclass remove small

* Configuration_file.rst: fix typo threshold -> thresholds

Co-authored-by: charleygros <charley.gros@gmail.com>

Co-authored-by: charleygros <charley.gros@gmail.com>
