### 2020-12-11 ee891a8
Add a Help section to the Sphinx docs (#571)
* Added README for getting started with the Sphinx docs

* Add help section to Sphinx docs

* remove README (out-of-scope of PR)
