### 2020-12-13 95328de
README.md: virtualenv venv-ivadomed --python=python3.6
