+++ README.md
@@ -16,7 +16,7 @@ The technical documentation is available [here](https://ivadomed.org).
 ``ivadomed`` requires Python >= 3.6 and < 3.9 as well as PyTorch == 1.5.0. We recommend working under a virtual environment, which could be set as follows:
 
 ```bash
-virtualenv venv-ivadomed --python=python3.6
+virtualenv venv-ivadomed
 source venv-ivadomed/bin/activate
 ```
 
