### 2020-12-14 f4491e1
update download_data so that args can be passed to main
