### 2020-12-14 #2868 @charleygros
Hi @lbugnon , We really appreciate that you took time to review our work!

> virtualenv venv-ivadomed --python=python3.6 gives an error (because i have installed python 3.7, I think It could be omited or stated to look into the OS

Thanks for raising this issue: it has been fixed in[ PR #579](https://github.com/ivadomed/ivadomed/pull/579) (omission of the Python version).

> Usage (as a demo) is not clear, for instance ivadomed ivadomed/config/config.json doesn't work. Adding -c option lead to error RuntimeError: The dataset_description.json file was not found. I think a "Please first look into the tutorials" would work

We agree with you @lbugnon: we added `-c` and directed the reader towards the tutorials for additional info, see [PR #578](https://github.com/ivadomed/ivadomed/pull/578).

> referneces are missing in the paper md

References are in `paper.bib` and correctly appear in the compiled document, see [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02868/joss.02868/10.21105.joss.02868.pdf) or via [Wheldom paper preview](Whedon paper preview service). Please let me know if I am missing something :-)
