### 2020-12-14 4b2a856
Fix the download_data argument parser
