### 2020-12-14 040e992
add get_arguments function to utils
