+++ ivadomed/utils.py
@@ -207,6 +207,10 @@ def check_exe(name):
     return None
 
 
+class ArgParseException(Exception):
+    pass
+
+
 def get_arguments(parser, args):
     """Get arguments from function input or command line.
 
@@ -216,10 +220,13 @@ def get_arguments(parser, args):
             should be formatted like this:
             ["-d", "SOME_ARG", "--model", "SOME_ARG"]
     """
-    if args:
-        args = parser.parse_args(args)
-    else:
-        args = parser.parse_args(args=None if sys.argv[1:] else ['--help'])
+    try:
+        if args:
+            args = parser.parse_args(args)
+        else:
+            args = parser.parse_args(args=None if sys.argv[1:] else ['--help'])
+    except SystemExit:
+        raise ArgParseException('Error parsing args')
     return args
 
 
