### 2020-12-14 a29e158
Create base testing file for CLI tests
