### 2020-12-14 945a089
Write tests for convert_to_onnx.py
