+++ testing/functional_tests/test_convert_to_onnx.py
@@ -1,0 +1,32 @@
+import logging
+import pytest
+import os
+from cli_base import download_test_data, __data_testing_dir__, remove_test_data
+from ivadomed.scripts import convert_to_onnx
+from ivadomed.utils import ArgParseException
+logger = logging.getLogger(__name__)
+
+
+def setup_function():
+    download_test_data()
+
+
+def test_convert_to_onnx():
+    model_path = os.path.join(__data_testing_dir__, 'spinegeneric_model.pt')
+    convert_to_onnx.main(args=['-m', f'{model_path}', '-d', '2'])
+    assert os.path.exists(os.path.join(__data_testing_dir__, 'spinegeneric_model.onnx'))
+
+
+def test_convert_to_onnx_no_model():
+    with pytest.raises(ArgParseException, match=r"Error parsing args"):
+        convert_to_onnx.main(args=['-d', '2'])
+
+
+def test_convert_to_onnx_no_dimension():
+    with pytest.raises(ArgParseException, match=r"Error parsing args"):
+        model_path = os.path.join(__data_testing_dir__, 'spinegeneric_model.pt')
+        convert_to_onnx.main(args=['-m', f'{model_path}'])
+
+
+def teardown_function():
+    remove_test_data()
