### 2020-12-15 f793ccc
add get_arguments to download_data
