### 2020-12-15 ffafc81
refactor cli_base to make download/remove dataset generic
