### 2020-12-15 b5bca37
refactor test_convert_to_onnx for dataset download/remove changes
