### 2020-12-15 4ac3d21
write CLI test for download_data.py
