### 2020-12-15 a5c1e42
WIP add multiple datasets on `config.json` bids_path entries will be be a list instead of a string
