### 2020-12-15 b517a9d
1.Took care of potential error when merging dataframes that have the same column with different DataTypes. 2. Generalized to any number of datasets
