### 2020-12-15 1b12505
Got rid of Neuropoly dependency
