+++ dev/merge_BIDS_datasets.py
@@ -1,5 +1,4 @@
 from distutils.dir_util import copy_tree
-from bids_neuropoly import bids
 import pandas as pd
 import json
 import glob
@@ -63,7 +62,8 @@ def main_run(argv):
         raise Exception("Less than 2 datasets selected. No need to merge")
     else:
         # Merge multiple .tsv files into the same dataframe
-        df_merged = bids.BIDS(path_folders[0]).participants.content
+        df_merged = pd.read_table(os.path.join(path_folders[0], 'participants.tsv'))
+
         # Convert to string to get rid of potential TypeError during merging within the same column
         df_merged = df_merged.astype(str)
 
@@ -73,7 +73,7 @@ def main_run(argv):
             json_merged = json.load(json_file)
 
         for iFolder in range(1, len(path_folders)):
-            df_next = bids.BIDS(path_folders[iFolder]).participants.content
+            df_next = pd.read_table(os.path.join(path_folders[iFolder], 'participants.tsv'))
             df_next = df_next.astype(str)
             # Merge the .tsv files (This keeps also non-overlapping fields)
             df_merged = pd.merge(left=df_merged, right=df_next, how='outer')
