### 2020-12-15 cb2cd37
Add tmp directory creation/removal for testing data
