### 2020-12-15 1b74d7f
Use tmp directory and fixtures instead of download_data in test_convert_to_onnx
