### 2020-12-15 6808444
Fix filtering when target_suffix is a nested list
