### 2020-12-15 400a1ec
Fix filtering when no contrast_lst is provided
