### 2020-12-16 403916d
WIP Working version of merging multiple BIDS datasets from selecting them from a list on config.json.
TODO: test on HEMIS
TODO: test Bounding box entries on: loader/loader.py
