+++ ivadomed/loader/loader.py
@@ -3,6 +3,8 @@ import random
 import nibabel as nib
 import numpy as np
 import torch
+import pandas as pd
+import os
 from bids_neuropoly import bids
 from torch.utils.data import Dataset
 from tqdm import tqdm
@@ -750,7 +752,7 @@ class BidsDataset(MRI2DSegmentationDataset):
     """ BIDS specific dataset loader.
 
     Args:
-        root_dir (str): Path to the BIDS dataset.
+        root_dirs (list): List of Paths to the BIDS datasets.
         subject_lst (list): Subject names list.
         target_suffix (list): List of suffixes for target masks.
         contrast_params (dict): Contains image contrasts related parameters.
@@ -778,11 +780,15 @@ class BidsDataset(MRI2DSegmentationDataset):
 
     """
 
-    def __init__(self, root_dir, subject_lst, target_suffix, contrast_params, slice_axis=2,
+    def __init__(self, root_BIDS_list, subject_lst, target_suffix, contrast_params, slice_axis=2,
                  cache=True, transform=None, metadata_choice=False, slice_filter_fn=None, roi_params=None,
                  multichannel=False, object_detection_params=None, task="segmentation", soft_gt=False):
 
-        self.bids_ds = bids.BIDS(root_dir)
+        #self.bids_ds = bids.BIDS(root_BIDS_list)
+        self.bids_ds = []
+        for BIDSFolder in root_BIDS_list:
+            self.bids_ds.append(bids.BIDS(BIDSFolder))
+
         self.roi_params = roi_params if roi_params is not None else {"suffix": None, "slice_filter_roi": None}
         self.soft_gt = soft_gt
         self.filename_pairs = []
@@ -790,7 +796,11 @@ class BidsDataset(MRI2DSegmentationDataset):
             self.metadata = {"FlipAngle": [], "RepetitionTime": [],
                              "EchoTime": [], "Manufacturer": []}
 
-        bids_subjects = [s for s in self.bids_ds.get_subjects() if s.record["subject_id"] in subject_lst]
+
+        # Append subjects from all BIDSdatasets into a list
+        bids_subjects = [s for s in self.bids_ds[0].get_subjects() if s.record["subject_id"] in subject_lst]
+        for iBIDSFolder in range(1, len(root_BIDS_list)):
+            bids_subjects = bids_subjects + [s for s in self.bids_ds[iBIDSFolder].get_subjects() if s.record["subject_id"] in subject_lst]
 
         # Create a list with the filenames for all contrasts and subjects
         subjects_tot = []
@@ -816,10 +826,21 @@ class BidsDataset(MRI2DSegmentationDataset):
                                                "roi_filename": None,
                                                "metadata": [None] * num_contrast} for subject in subject_lst}
 
+
+
+        ###############################################################
+
+        # TAKE CARE OF BOUNDING BOX AFTER CHANGES - WARNING - TEST THIS
+
+
         bounding_box_dict = imed_obj_detect.load_bounding_boxes(object_detection_params,
-                                                                self.bids_ds.get_subjects(),
-                                                                slice_axis,
-                                                                contrast_params["contrast_lst"])
+                                                                    self.bids_ds[0].get_subjects(),
+                                                                    slice_axis,
+                                                                    contrast_params["contrast_lst"])
+
+
+        ###############################################################
+
 
         for subject in tqdm(bids_subjects, desc="Loading dataset"):
             if subject.record["modality"] in contrast_params["contrast_lst"]:
@@ -865,19 +886,43 @@ class BidsDataset(MRI2DSegmentationDataset):
                 # add contrast to metadata
                 metadata['contrast'] = subject.record["modality"]
 
+
+
[…]

[…]
