### 2020-12-16 3cde87b
Took care of problematic encoding in some participants.tsv. In case more errors occur we might need to try several encodings until one works
