### 2020-12-16 80e2081
Add test_loader.py script to testing/
