### 2020-12-16 b8d941c
Create stubs for functional tests
