### 2020-12-16 b9499ad
Create test for automate_training - need to fix config files
