### 2020-12-16 15d0164
Make HDF5Dataset class return all modalities
