### 2020-12-16 44d2bf0
Fix shape error for multi-contrast HeMIS usage
