### 2020-12-16 8c2b3d5
refactor model parameters in main.py into get_model_parameters
