### 2020-12-16 9f4f238
refactor loader parameters in main.py into get_loader_parameters
