### 2020-12-16 ee7ac48
add more lines to get_model_parameters
