### 2020-12-16 70ba0cf
refactor training and threshold analysis in main
