### 2020-12-17 2c35e71
Took care of the possibility that the same subject exists in more than one dataset. Only the files within the first BIDS dataset will be used
