### 2020-12-17 38a515d
WIP - Still need to do the testing with bounding boxes etc.
Corrections based on Andreanne's suggestions.
A string for bids_path can be assigned at config.json
