### 2020-12-17 7d2c3c7
add args to automate_training
