### 2020-12-17 4f20363
download spinegeneric dataset for automate_training
