+++ testing/functional_tests/test_automate_training.py
@@ -1,16 +1,17 @@
 import logging
 import os
-from cli_base import remove_tmp_dir, __tmp_dir__, create_tmp_dir
+from cli_base import remove_tmp_dir, __tmp_dir__, download_dataset, remove_dataset, create_tmp_dir
 from ivadomed.scripts import automate_training
 logger = logging.getLogger(__name__)
 
 
 def setup_function():
     create_tmp_dir()
+    download_dataset("data_example_spinegeneric")
 
 
 def test_automate_training():
-    config_file = os.path.join(__tmp_dir__, 'automate_training_config.json')
+    config_file = os.path.join(__tmp_dir__, 'automate_training_config_test.json')
     param_file = os.path.join(__tmp_dir__,
                               'automate_training_hyperparameter_opt.json')
     automate_training.main(args=[
@@ -20,4 +21,5 @@ def test_automate_training():
 
 
 def teardown_function():
+    remove_dataset("data_example_spinegeneric")
     remove_tmp_dir()
