### 2020-12-17 58a2b43
refactor testing into run_testing in main
