### 2020-12-17 23ae3c2
Add docstring to save_config_file
