### 2020-12-17 7e38bba
refactor testing params into get_testing_parameters
