### 2020-12-17 7254f16
refactor segment code into run_segmentation
