+++ ivadomed/loader/loader.py
@@ -79,7 +79,6 @@ def load_dataset(data_list, bids_path, transforms_params, model_params, target_s
                                               slice_axis=imed_utils.AXIS_DCT[slice_axis],
                                               transform=tranform_lst,
                                               metadata_choice=metadata_type,
-                                              multichannel=multichannel,
                                               slice_filter_fn=imed_loader_utils.SliceFilter(**slice_filter_params,
                                                                                             device=device,
                                                                                             cuda_available=cuda_available),
