### 2020-12-17 4e50327
move downloaded datasets into tmp dir for testing
