### 2020-12-17 1147e19
test_HeMIS: fix bug
