### 2020-12-17 b16950c
Added support for selecting the --metric to be displayed (dice_class0 is default) and additionally an optional --metadata selection (2 inputs - (1) column of metadata in participants.tsv and (2) string to match)
