+++ ivadomed/scripts/visualize_and_compare_testing_models.py
@@ -1,0 +1,172 @@
+import matplotlib
+matplotlib.use('TkAgg')  # This is needed for plotting through a CLI call
+import matplotlib.pyplot as plt
+import pandas as pd
+import os
+import numpy as np
+import itertools
+import seaborn as sns
+from scipy.stats import ks_2samp
+import argparse
+import sys
+
+
+# This function allows violinplots visualization of multiple evaluation models simultaneously and performs a
+# Kolmogorov–Smirnov significance test between each combination of models.
+# If only one model is selected as input, only the Violinplot will be presented (no test will be superimposed)
+
+# Inputs:
+# --listfolders: list of folders that contain the logs - space separated
+# --metric: column of "results_eval/evaluation_3Dmetrics.csv" to be used on the plots e.g. dice_class0
+# --metadata (optional): 2 elements - (1) column of the participants.tsv metadata so only subjects that belong to that
+#                        category will be used and (2) string to be matched e.g. pathology ms
+
+# Example calls from terminal:
+# python3 visualize_and_compare_testing_models.py --listfolders ~/logs/logs_NO_FILM_sctUsers ~/logs/logs_onlyT1w
+# or
+# python3 visualize_and_compare_testing_models.py --listfolders /home/nas/Desktop/logs/logs_*
+#           --metric dice_class0 --metadata pathology ms
+
+
+# Konstantinos Nasiotis 2020
+
+
+# Dependency: sudo apt-get install python3-tk
+# - needed for matplotlib visualization through a CLI call
+# ----------------------------------------------------------------------------------------------------------------------#
+
+def main(argv):
+    # defined command line options
+    # this also generates --help and error handling
+    CLI = argparse.ArgumentParser()
+    CLI.add_argument(
+        "--listfolders",
+        nargs="*",  # 0 or more values expected => creates a list
+        type=str,
+        default=[],  # default if nothing is provided
+    )
+    CLI.add_argument(
+        "--metric",
+        nargs=1,
+        type=str,
+        default=["dice_class0"],
+    )
+    CLI.add_argument(
+        "--metadata",
+        nargs=2,
+        type=str,
+        default=[],
+    )
+
+    # parse the command line
+    args = CLI.parse_args()
+    args.metric = args.metric[0]
+
+    # access CLI options
+    print("listfolders: %r" % args.listfolders)
+    print("metric: %r" % args.metric)
+    if args.metadata != []:
+        print("metadata: %r" % args.metadata)
+
+    # Get the list
+    logFoldersToCompare = args.listfolders
+    # Do a quick check that all the required files are present
+    for folder in logFoldersToCompare:
+        if not os.path.exists(os.path.join(folder, 'results_eval', 'evaluation_3Dmetrics.csv')):
+            print('evaluation_3Dmetrics.csv file is not present within ' + os.path.join(folder, 'results_eval'))
+            raise Exception('evaluation_3Dmetrics.csv missing')
+        if not os.path.exists(os.path.join(folder, 'participants.tsv')):
+            print('participants.tsv file is not present within ' + folder)
+            raise Exception('participants.tsv missing')
+
+    if len(logFoldersToCompare) < 1:
+        raise Exception('No folders were selected - Nothing to show')
+
+    columnNames = ["EvaluationModel", args.metric]
+    df = pd.DataFrame([], columns=columnNames)
+
+    for folder in logFoldersToCompare:
+        result = pd.read_csv(os.path.join(folder, 'results_eval', 'evaluation_3Dmetrics.csv'))
+
+        if args.metadata:
+            participant_metadata = pd.read_table(os.path.join(folder, 'participants.tsv'), encoding="ISO-8859-1")
[…]
