### 2020-12-18 fab21b0
add functional testing release to download_data
