### 2020-12-18 29a8359
replace dataset in testing automate_training
