### 2020-12-18 1f58480
add output_dir to automate_training, add asserts to test
