### 2020-12-18 bcdf726
remove spinegeneric_model.pt file and use file from data_functional_testing archinve
