+++ testing/functional_tests/test_convert_to_onnx.py
@@ -1,20 +1,23 @@
 import logging
 import pytest
 import os
-from cli_base import remove_tmp_dir, __tmp_dir__, create_tmp_dir
+from cli_base import remove_tmp_dir, __tmp_dir__, create_tmp_dir, download_dataset
 from ivadomed.scripts import convert_to_onnx
 from ivadomed.utils import ArgParseException
 logger = logging.getLogger(__name__)
 
+__model_path__ = os.path.join(__tmp_dir__, 'data_functional_testing', 'spinegeneric_model.pt')
+
 
 def setup_function():
     create_tmp_dir()
+    download_dataset("data_functional_testing")
 
 
 def test_convert_to_onnx():
-    model_path = os.path.join(__tmp_dir__, 'spinegeneric_model.pt')
-    convert_to_onnx.main(args=['-m', f'{model_path}', '-d', '2'])
-    assert os.path.exists(os.path.join(__tmp_dir__, 'spinegeneric_model.onnx'))
+    convert_to_onnx.main(args=['-m', f'{__model_path__}', '-d', '2'])
+    assert os.path.exists(os.path.join(__tmp_dir__, 'data_functional_testing',
+                                       'spinegeneric_model.onnx'))
 
 
 def test_convert_to_onnx_no_model():
@@ -24,8 +27,7 @@ def test_convert_to_onnx_no_model():
 
 def test_convert_to_onnx_no_dimension():
     with pytest.raises(ArgParseException, match=r"Error parsing args"):
-        model_path = os.path.join(__tmp_dir__, 'spinegeneric_model.pt')
-        convert_to_onnx.main(args=['-m', f'{model_path}'])
+        convert_to_onnx.main(args=['-m', f'{__model_path__}'])
 
 
 def teardown_function():
