### 2020-12-18 4118807
move test_script to functional_tests
