+++ testing/functional_tests/test_script.py
@@ -1,0 +1,394 @@
+import json
+import os
+import subprocess
+from csv import writer
+import shutil
+
+import nibabel as nib
+import pytest
+import torch
+
+from ivadomed import config_manager as imed_config_manager
+import ivadomed.models as imed_models
+from ivadomed import main as imed
+from ivadomed import utils as imed_utils
+from ivadomed import inference as imed_inference
+
+
+def test_download_data():
+    command = "ivadomed_download_data -d t2_tumor"
+    subprocess.check_output(command, shell=True)
+    command = "ivadomed_download_data -d findcord_tumor"
+    subprocess.check_output(command, shell=True)
+    command = "ivadomed_download_data -d data_testing -o output -k 1"
+    subprocess.check_output(command, shell=True)
+
+
+def test_create_segment_file():
+    command = "cp testing_data/model_config_test.json testing_data/model_config_segment.json"
+    subprocess.check_output(command, shell=True)
+    os.makedirs("testing_script", exist_ok=True)
+    command = "scp -r findcord_tumor testing_script"
+    subprocess.check_output(command, shell=True)
+    file_conf = open("testing_data/model_config_segment.json", "r")
+    initial_config = json.load(file_conf)
+    file_conf.close()
+    file_conf = open("testing_data/model_config_segment.json", "w")
+    initial_config['command'] = "segment"
+    initial_config['model_name'] = "findcord_tumor"
+    initial_config['postprocessing'] = {}
+    initial_config['split_dataset']['test_fraction'] = 1.0
+
+    json.dump(initial_config, file_conf)
+
+
+def test_segment():
+    subprocess.check_output(["ivadomed -c testing_data/model_config_segment.json"], shell=True)
+    shutil.rmtree(os.path.join('testing_script', 'pred_masks'))
+
+
+def test_onnx_conversion():
+    # testing convert to onnx
+    subprocess.check_output("ivadomed_convert_to_onnx -m testing_data/model_unet_test.pt -d 2", shell=True)
+
+
+def test_prepare_dataset_vertebral_labeling():
+    # testing prepare_dataset_vertebral_labeling
+    subprocess.check_output("ivadomed_prepare_dataset_vertebral_labeling -p testing_data/ -s _T2w -a 3", shell=True)
+
+
+def test_visualize_transform():
+    # testing visualize_transform
+    command = "ivadomed_visualize_transforms -i testing_data/sub-unf01/anat/sub-unf01_T1w.nii.gz -n " + \
+              "2 -c testing_data/model_config.json " + \
+              "-r testing_data/derivatives/labels/sub-test001/anat/sub-unf01_T1w_seg-manual.nii.gz -o visuzalize_test"
+    subprocess.check_output(command, shell=True)
+
+
+def test_extract_small():
+    # testing extract_small_dataset
+    subprocess.check_output("ivadomed_extract_small_dataset -i testing_data/ -o small_dataset/test_script/ -n 1 -c T2w,"
+                            "T1w -d 1", shell=True)
+
+
+def test_compare_model():
+    # testing compare_model
+    command = "ivadomed_compare_models -df testing_data/temporary_results.csv -n 2 -o output_test.csv"
+    subprocess.check_output(command, shell=True)
+
+
+def test_creation_dataset():
+    # Add new file as needed (no empty test/validation)
+    # create empty directory for our new files
+    os.makedirs("testing_data/sub-test002/anat/", exist_ok=True)
+    os.makedirs("testing_data/sub-test003/anat/", exist_ok=True)
+    os.makedirs("testing_data/derivatives/labels/sub-test002/anat/", exist_ok=True)
+    os.makedirs("testing_data/derivatives/labels/sub-test003/anat/", exist_ok=True)
+
[…]
