### 2020-12-18 01fefc3
loader: keep only subjects with GT for training
