+++ testing/test_HeMIS.py
@@ -80,7 +80,7 @@ def test_HeMIS(p=0.0001):
                                           metadata_choice=False,
                                           dim=2,
                                           slice_filter_fn=imed_loader_utils.SliceFilter(filter_empty_input=True,
-                                                                                 filter_empty_mask=True),
+                                                                                        filter_empty_mask=True),
                                           roi_params=roi_params)
 
     dataset.load_into_ram(['T1w', 'T2w', 'T2star'])
