### 2020-12-18 eb04460
adaptive: store slices by contrast
