### 2020-12-18 d77243d
Bids_to_hdf5: make contrasts mandatory param
