### 2020-12-18 873440a
Move output_dir from config file and into parser args for automate_training
