### 2020-12-19 9d134f2
Code refactoring to match project's guidelines. Added screenshot of the Violinplots
