+++ ivadomed/scripts/visualize_and_compare_testing_models.py
@@ -1,5 +1,13 @@
+###########################################################################################################
+#
+# This script makes a figure with violinplots and significance values to compare between models
+#
+#        python3 visualize_and_compare_testing_models.py --logfolders path/to/logfolder1 path/to/logfolder2
+#                              --metric metric/to/use --metadata metadata/label string/to/match
+###########################################################################################################
+
+
 import matplotlib
-matplotlib.use('TkAgg')  # This is needed for plotting through a CLI call
 import matplotlib.pyplot as plt
 import pandas as pd
 import os
@@ -7,71 +15,65 @@ import numpy as np
 import itertools
 import seaborn as sns
 from scipy.stats import ks_2samp
+from ivadomed.utils import init_ivadomed
 import argparse
-import sys
-
-
-# This function allows violinplots visualization of multiple evaluation models simultaneously and performs a
-# Kolmogorov–Smirnov significance test between each combination of models.
-# If only one model is selected as input, only the Violinplot will be presented (no test will be superimposed)
-
-# Inputs:
-# --listfolders: list of folders that contain the logs - space separated
-# --metric: column of "results_eval/evaluation_3Dmetrics.csv" to be used on the plots e.g. dice_class0
-# --metadata (optional): 2 elements - (1) column of the participants.tsv metadata so only subjects that belong to that
-#                        category will be used and (2) string to be matched e.g. pathology ms
-
-# Example calls from terminal:
-# python3 visualize_and_compare_testing_models.py --listfolders ~/logs/logs_NO_FILM_sctUsers ~/logs/logs_onlyT1w
-# or
-# python3 visualize_and_compare_testing_models.py --listfolders /home/nas/Desktop/logs/logs_*
-#           --metric dice_class0 --metadata pathology ms
-
+matplotlib.use('TkAgg')  # This is needed for plotting through a CLI call
 
 # Konstantinos Nasiotis 2020
-
-
+#
 # Dependency: sudo apt-get install python3-tk
 # - needed for matplotlib visualization through a CLI call
 # ----------------------------------------------------------------------------------------------------------------------#
 
-def main(argv):
-    # defined command line options
-    # this also generates --help and error handling
-    CLI = argparse.ArgumentParser()
-    CLI.add_argument(
-        "--listfolders",
-        nargs="*",  # 0 or more values expected => creates a list
-        type=str,
-        default=[],  # default if nothing is provided
-    )
-    CLI.add_argument(
-        "--metric",
-        nargs=1,
-        type=str,
-        default=["dice_class0"],
-    )
-    CLI.add_argument(
-        "--metadata",
-        nargs=2,
-        type=str,
-        default=[],
-    )
 
-    # parse the command line
-    args = CLI.parse_args()
-    args.metric = args.metric[0]
+def get_parser():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--logfolders", required=True, nargs="*", dest="logfolders",
+                        help="List of log folders from different models.")
+    parser.add_argument("--metric", default='dice_class0', nargs=1, type=str, dest="metric",
+                        help="Metric from evaluation_3Dmetrics.csv to base the plots on.")
+    parser.add_argument("--metadata", required=False,  nargs=2, type=str, dest="metadata",
+                        help="Selection based on metadata from participants.tsv: (1) Label from column (2) string to match")
+    return parser
+
+
+def visualize_and_compare_models(logfolders, metric, metadata):
[…]
