### 2020-12-19 b3b28e5
change __tmp_dir__ to be relative to repo, not absolute
