### 2020-12-21 ac45784
Move test_loader.py to folder unit_tests
