### 2020-12-21 5565d4f
Remove test_loader.py from testing (moved to unit_tests)
