### 2020-12-21 8b8594b
Update comments in ivadomed/loader/utils.py
Co-authored-by: andreanne-lemay <49137243+andreanne-lemay@users.noreply.github.com>
