### 2020-12-21 8d4bd9b
add helper file unit_base for unit tests
