### 2020-12-22 e037328
loader/utils.py: Add "get_subdatasets_subjects_list_new" function
