### 2020-12-22 91e5b32
loader/utils.py: Add "get_new_subject_split_new" function
