### 2020-12-22 bc483fd
loader/utils.py: Add "split_dataset_new" function
