### 2020-12-22 717e529
dev/df_new_loader.py: UPdate script to test new splitting method
