### 2020-12-22 b1e08d4
Update docstrings
