### 2020-12-22 1c777ec
refactor adaptative code to open/close hdf5 files properly
