+++ testing/unit_tests/test_adaptative.py
@@ -1,5 +1,5 @@
 import os
-
+import h5py
 import torch
 from torch.utils.data import DataLoader
 
@@ -38,17 +38,17 @@ def test_hdf5():
 
     roi_params = {"suffix": "_seg-manual", "slice_filter_roi": None}
 
-    hdf5_file = imed_adaptative.Bids_to_hdf5(PATH_BIDS,
-                                             subject_lst=train_lst,
-                                             hdf5_name='testing_data/mytestfile.hdf5',
-                                             target_suffix=["_lesion-manual"],
-                                             roi_params=roi_params,
-                                             contrast_lst=['T1w', 'T2w', 'T2star'],
-                                             metadata_choice="contrast",
-                                             transform=transform_lst,
-                                             contrast_balance={},
-                                             slice_axis=2,
-                                             slice_filter_fn=imed_loader_utils.SliceFilter(filter_empty_input=True,
+    bids_to_hdf5 = imed_adaptative.Bids_to_hdf5(PATH_BIDS,
+                                                subject_lst=train_lst,
+                                                hdf5_path='testing_data/mytestfile.hdf5',
+                                                target_suffix=["_lesion-manual"],
+                                                roi_params=roi_params,
+                                                contrast_lst=['T1w', 'T2w', 'T2star'],
+                                                metadata_choice="contrast",
+                                                transform=transform_lst,
+                                                contrast_balance={},
+                                                slice_axis=2,
+                                                slice_filter_fn=imed_loader_utils.SliceFilter(filter_empty_input=True,
                                                                                     filter_empty_mask=True))
 
     # Checking architecture
@@ -60,90 +60,91 @@ def test_hdf5():
             print("    %s: %s" % (key, val))
 
     print('\n[INFO]: HDF5 architecture:')
-    hdf5_file.hdf5_file.visititems(print_attrs)
-    print('\n[INFO]: HDF5 file successfully generated.')
-    print('[INFO]: Generating dataframe ...\n')
+    with h5py.File(bids_to_hdf5.hdf5_path, "a") as hdf5_file:
+        hdf5_file.visititems(print_attrs)
+        print('\n[INFO]: HDF5 file successfully generated.')
+        print('[INFO]: Generating dataframe ...\n')
 
-    df = imed_adaptative.Dataframe(hdf5=hdf5_file.hdf5_file,
-                                   contrasts=['T1w', 'T2w', 'T2star'],
-                                   path='testing_data/hdf5.csv',
-                                   target_suffix=['T1w', 'T2w', 'T2star'],
-                                   roi_suffix=['T1w', 'T2w', 'T2star'],
-                                   dim=2,
-                                   filter_slices=True)
+        df = imed_adaptative.Dataframe(hdf5_file=hdf5_file,
+                                       contrasts=['T1w', 'T2w', 'T2star'],
+                                       path='testing_data/hdf5.csv',
+                                       target_suffix=['T1w', 'T2w', 'T2star'],
+                                       roi_suffix=['T1w', 'T2w', 'T2star'],
+                                       dim=2,
+                                       filter_slices=True)
 
-    print(df.df)
+        print(df.df)
 
-    print('\n[INFO]: Dataframe successfully generated. ')
-    print('[INFO]: Creating dataset ...\n')
+        print('\n[INFO]: Dataframe successfully generated. ')
+        print('[INFO]: Creating dataset ...\n')
 
-    model_params = {
-            "name": "HeMISUnet",
-            "dropout_rate": 0.3,
-            "bn_momentum": 0.9,
-            "depth": 2,
-            "in_channel": 1,
-            "out_channel": 1,
-            "missing_probability": 0.00001,
[…]
