+++ ivadomed/loader/adaptative.py
@@ -17,11 +17,12 @@ from ivadomed.object_detection import utils as imed_obj_detect
 class Dataframe:
     """
     This class aims to create a dataset using an HDF5 file, which can be used by an adapative loader
-    to perform curriculum learning, Active Learning or any other strategy that needs to load samples in a specific way.
+    to perform curriculum learning, Active Learning or any other strategy that needs to load samples
+    in a specific way.
     It works on RAM or on the fly and can be saved for later.
 
     Args:
-        hdf5 (hdf5): hdf5 file containing dataset information
+        hdf5_file (hdf5): hdf5 file containing dataset information
         contrasts (list of str): List of the contrasts of interest.
         path (str): Dataframe path.
         target_suffix (list of str): List of suffix of targetted structures.
@@ -36,7 +37,7 @@ class Dataframe:
         df (pd.Dataframe): Dataframe containing dataset information
     """
 
-    def __init__(self, hdf5, contrasts, path, target_suffix=None, roi_suffix=None,
+    def __init__(self, hdf5_file, contrasts, path, target_suffix=None, roi_suffix=None,
                  filter_slices=False, dim=2):
         # Number of dimension
         self.dim = dim
@@ -62,7 +63,7 @@ class Dataframe:
         if os.path.exists(path):
             self.load_dataframe(path)
         else:
-            self.create_df(hdf5)
+            self.create_df(hdf5_file)
 
     def shuffle(self):
         """Shuffle the whole data frame."""
@@ -92,11 +93,11 @@ class Dataframe:
         except FileNotFoundError:
             print("Wrong path.")
 
-    def create_df(self, hdf5):
+    def create_df(self, hdf5_file):
         """Generate the Data frame using the hdf5 file.
 
         Args:
-            hdf5 (hdf5): File containing dataset information
+            hdf5_file (hdf5): File containing dataset information
         """
         # Template of a line
         empty_line = {col: 'None' for col in self.contrasts}
@@ -106,11 +107,11 @@ class Dataframe:
         col_names = [col for col in empty_line.keys()]
         col_names.append('Subjects')
         df = pd.DataFrame(columns=col_names)
-        print(hdf5.attrs['patients_id'])
+        print(hdf5_file.attrs['patients_id'])
         # Filling the data frame
-        for subject in hdf5.attrs['patients_id']:
+        for subject in hdf5_file.attrs['patients_id']:
             # Getting the Group the corresponding patient
-            grp = hdf5[subject]
+            grp = hdf5_file[subject]
             line = copy.deepcopy(empty_line)
             line['Subjects'] = subject
 
@@ -180,7 +181,7 @@ class Bids_to_hdf5:
         target_suffix (list): List of suffixes for target masks.
         roi_params (dict): Dictionary containing parameters related to ROI image processing.
         contrast_lst (list): List of the contrasts.
-        hdf5_name (str): Path and name of the hdf5 file.
+        hdf5_path (str): Path and name of the hdf5 file.
         contrast_balance (dict): Dictionary controlling image contrasts balance.
         slice_axis (int): Indicates the axis used to extract slices: "axial": 2, "sagittal": 0, "coronal": 1.
         metadata_choice (str): Choice between "mri_params", "contrasts", None or False, related to FiLM.
@@ -191,7 +192,7 @@ class Bids_to_hdf5:
     Attributes:
         bids_ds (BIDS): BIDS dataset.
         dt (dtype): hdf5 special dtype.
-        hdf5_file (hdf5): hdf5 file containing dataset information.
+        hdf5_path (str): path to hdf5 file containing dataset information.
         filename_pairs (list): A list of tuples in the format (input filename list containing all modalities,ground \
             truth filename, ROI filename, metadata).
[…]
