### 2020-12-22 1ea9603
rename hdf5/bids classes to be PEP8 compliant
