### 2020-12-22 04389b4
add pytest ordering, update test_HeMIS to delete files after test
