### 2020-12-22 4f1f020
rename hdf5_path to path_hdf5 to be consistent with naming conventions
