### 2020-12-23 c85a387
Add new filter to SliceFilter class
