+++ ivadomed/loader/utils.py
@@ -465,17 +465,23 @@ class SliceFilter(object):
     Args:
         filter_empty_mask (bool): If True, samples where all voxel labels are zeros are discarded.
         filter_empty_input (bool): If True, samples where all voxel intensities are zeros are discarded.
+        filter_absent_mask (bool): If True, samples where all voxel labels in one or more masks are discarded.
 
     Attributes:
         filter_empty_mask (bool): If True, samples where all voxel labels are zeros are discarded.
         filter_empty_input (bool): If True, samples where all voxel intensities are zeros are discarded.
+        filter_absent_mask (bool): If True, samples where all voxel labels in one or more masks are discarded.
+
     """
 
     def __init__(self, filter_empty_mask=True,
                  filter_empty_input=True,
-                 filter_classification=False, classifier_path=None, device=None, cuda_available=None):
+                 filter_classification=False,
+                 filter_absent_mask=False,
+                 classifier_path=None, device=None, cuda_available=None):
         self.filter_empty_mask = filter_empty_mask
         self.filter_empty_input = filter_empty_input
+        self.filter_absent_mask = filter_absent_mask
         self.filter_classification = filter_classification
         self.device = device
         self.cuda_available = cuda_available
@@ -490,7 +496,13 @@ class SliceFilter(object):
         input_data, gt_data = sample['input'], sample['gt']
 
         if self.filter_empty_mask:
+            # Filter slices that do not have ANY ground truth (i.e. all masks are empty)
             if not np.any(gt_data):
+                return False
+
+        if self.filter_absent_mask:
+            # Filter slices that have absent masks (i.e. one or more masks are empty)
+            if not np.all([np.any(mask) for mask in gt_data]):
                 return False
 
         if self.filter_empty_input:
