### 2020-12-23 82ed360
Fix filtering of dataframe when roi_suffix is provided
