### 2020-12-23 2fd8f89
Update bids dataframe test with roi_suffix param
