+++ testing/unit_tests/test_loader.py
@@ -12,6 +12,10 @@ from ivadomed.loader import utils as imed_loader_utils
     "bids_config": "ivadomed/config/config_bids.json",
     "target_suffix": [["_seg-myelin-manual", "_seg-axon-manual"]],
     "extensions": [".png"],
+    "roi_params": {
+        "suffix": None,
+        "slice_filter_roi": None
+    },
     "contrast_params": {
         "training_validation": [],
         "testing": [],
@@ -40,6 +44,10 @@ def test_bids_df_microscopy_png(loader_parameters):
     "bids_path": "testing_data",
     "target_suffix": ["_seg-manual"],
     "extensions": [],
+    "roi_params": {
+        "suffix": None,
+        "slice_filter_roi": None
+    },
     "contrast_params": {
         "training_validation": ["T1w", "T2w"],
         "testing": [],
@@ -49,6 +57,7 @@ def test_bids_df_anat(loader_parameters):
     # Test for MRI anat nii.gz file format
     # Test for when no file extensions are provided
     # Test for multiple target_suffix
+    # TODO: modify test and "df_ref.csv" file in data-testing dataset to test behavior when "roi_suffix" is not None
     loader_params = loader_parameters
     loader_params["contrast_params"]["contrast_lst"] = loader_params["contrast_params"]["training_validation"]
     bids_path = loader_params["bids_path"]
