### 2020-12-23 76d8375
Split dataset returns lists according to data_type
