### 2020-12-23 d877a2b
Update bids dataframe test
