### 2020-12-25 474002d
Update data.rst
Made the BIDS example more general for the audience
