### 2020-12-26 94ebc54
Revert "Fix shape error for multi-contrast HeMIS usage"
This reverts commit 44d2bf04a71bc4a0ec8a24ed1132de6d3e0f2d6a.
