### 2020-12-27 d22b6d4
Save entire crop params dict into HDF5 metadata attributes
HDF5 doesn't allow dictionaries to be assigned as metadata attributes,
so as a work-around, key:value pairs from the dict are now stored as
grp[key].attrs["crop_params/<key] = value, and later retrieved properly
properly in `HDF5Dataset.__getitem__`.
