### 2020-12-28 0726e0c
Save preresample_shape in HDF5 file
