### 2020-12-28 1d9211a
Add documentation for SliceFilter class
