### 2021-01-02 abe293f
Make final activation selectable for HeMIS
