### 2021-01-02 417511f
Output argmax'd predictions in one plane
