### 2021-01-04 0ebe217
Fix loader/adaptative.py code with reading/writing HDF5 files (#592)
* refactor adaptative code to open/close hdf5 files properly

* rename hdf5/bids classes to be PEP8 compliant

* rename hdf5_path to path_hdf5 to be consistent with naming conventions
