### 2021-01-04 7f031ad
clean up cli_base
