### 2021-01-05 30f826f
Explicit what sub-01 meant
Addressing https://github.com/ivadomed/ivadomed/pull/597#pullrequestreview-561450940
