### 2021-01-06 8988dfb
Make final activation selection for HeMIS optional
