### 2021-01-06 a7de5c4
Add 'HeMIS' keyword to if statements
