+++ ivadomed/main.py
@@ -16,7 +16,7 @@ from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader, f
 cudnn.benchmark = True
 
 # List of not-default available models i.e. different from Unet
-MODEL_LIST = ['UNet3D', 'HeMISUnet', 'FiLMedUnet', 'resnet18', 'densenet121', 'Countception']
+MODEL_LIST = ['UNet3D', 'HeMIS', 'HeMISUnet', 'FiLMedUnet', 'resnet18', 'densenet121', 'Countception']
 
 
 def get_parser():
