### 2021-01-07 ivadomed/ivadomed#604 @ahill187
Add compatibility for Windows
We would like to add compatibility for `Windows` in `ivadomed`. Currently we are only `Linux` and `Mac` compatible.

### Motivation for the feature
In this review, some of the reviewers mentioned that they had issues trying to install `ivadomed` on `Windows`: https://github.com/openjournals/joss-reviews/issues/2868#issuecomment-752687437

### Description of the feature
To address the issue in the review, we will add to our documentation that we do not currently support Windows. However, it would be nice to support Windows in the future.
