### 2021-01-08 4152f99
add test for compare_models
