+++ docs/source/installation.rst
@@ -9,6 +9,15 @@ working under a virtual environment, which could be set as follows:
     virtualenv venv-ivadomed
     source venv-ivadomed/bin/activate
 
+.. warning:: 
+   If the default Python version installed in your system does not fit the version requirements, you might need to specify a version of Python associated with your virtual environment:
+   
+   ::
+   
+     virtualenv venv-ivadomed --python=python3.6
+
+
+
 Install from release (recommended)
 ----------------------------------
 
