### 2021-01-10 7cd7850
Updated "bids_path" for consistency
I didn't notice that the convention was to use the same name as the field name in the JSON config file for the example FSLeyes command
