### 2021-01-10 086cec3
Fixed formatting issue with bullet points
