### 2021-01-10 9f31b36
Fixed formatting issue with bullet points
