+++ paper.bib
@@ -44,11 +44,15 @@
 	file = {PubMed Central Full Text PDF:C\:\\Users\\Lucas\\Zotero\\storage\\BEKHXIPR\\Gros et al. - 2019 - Automatic segmentation of the spinal cord and intr.pdf:application/pdf}
 }
 
-@article{paszke2017automatic,
-  title={Automatic differentiation in PyTorch},
-  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  journal = {NeurIPS Workshop},
-  year={2017}
+@incollection{NEURIPS2019_9015,
+	title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+	author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
+	booktitle = {Advances in Neural Information Processing Systems 32},
+	editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
+	pages = {8024--8035},
+	year = {2019},
+	publisher = {Curran Associates, Inc.},
+	url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
 
 @article{kim_deep_2019,
