### 2021-01-10 #2868 @charleygros
> The [following reference](https://github.com/ivadomed/ivadomed/blob/master/paper.bib#L47), which is supposed to reference Pytorch, whould probably be replaced by this one

Thanks for your suggestion @jcohenadad. Addressed in this [Pull Request](https://github.com/ivadomed/ivadomed/pull/616).
