### 2021-01-11 #2868 @charleygros
Many thanks for taking the time to review our work and providing valuable feedback @NMontanaBrown!

> I tried this with various python versions to no avail. Overall, this seems to be related to WindowsOS, as the installation proceeds fine on other OS. I would suggest updating your RTD installation section and the README to state installation works as described for MacOS and Linux, but not currently supported for Windows. It may be something you want to fix or not - but this will clarify upfront to users that installation works as intended except on Windows and avoid head scratching at the first stage.

Thank you for raising this point! We followed your advice by:
- Opening a Feature Request to add compatibility to Windows → https://github.com/ivadomed/ivadomed/issues/604
- Adding  explanation for Windows incompatibility in installation docs → https://github.com/ivadomed/ivadomed/pull/605
- Specify Python version in setup.py → https://github.com/ivadomed/ivadomed/pull/603

> In your table you include some examples of registration packages that use deep learning, yet ivadomed does not yet cover these functionalities (yet (?)). Are you planning on covering this use case in the future? Additionally, would you consider explaining this in your paper, and adding [DeepReg](https://doi.org/10.21105/joss.02705) to your table/citing it?

Thank you for this relevant suggestion! Indeed, `ivadomed` does not cover registration (yet). We have added a column for `DeepReg`, see [here](https://docs.google.com/spreadsheets/d/1_MydnHnlOAuYzJ9QBCvPC9Jq2xUmPWI-XttTfcdtW2Y/edit?usp=sharing) (Note: it will appear [here](https://ivadomed.org/en/latest/purpose.html) after the release). Please feel free to comment/suggest modifications if we missed some features of `DeepReg`.
