### 2021-01-11 0858b39
purpose.rst: typo
