+++ docs/source/purpose.rst
@@ -31,7 +31,7 @@ understand that the field is rapidly evolving, and that this table might reflect
 inconsistencies, please let us know by `opening an issue <https://github.com/ivadomed/ivadomed/issues>`_.
 
 ..
-  If you wish to modify the csv tbale please modify https://docs.google.com/spreadsheets/d/1_MydnHnlOAuYzJ9QBCvPC9Jq2xUmPWI-XttTfcdtW2Y/edit#gid=0
+  If you wish to modify the csv table please modify https://docs.google.com/spreadsheets/d/1_MydnHnlOAuYzJ9QBCvPC9Jq2xUmPWI-XttTfcdtW2Y/edit#gid=0
 
 .. csv-table::
    :file: comparison_other_projects_table.csv
