### 2021-01-11 #2868 @NMontanaBrown
> Many thanks for taking the time to review our work and providing valuable feedback @NMontanaBrown!
>
> > I tried this with various python versions to no avail. Overall, this seems to be related to WindowsOS, as the installation proceeds fine on other OS. I would suggest updating your RTD installation section and the README to state installation works as described for MacOS and Linux, but not currently supported for Windows. It may be something you want to fix or not - but this will clarify upfront to users that installation works as intended except on Windows and avoid head scratching at the first stage.
>
> Thank you for raising this point! We followed your advice by:
>
> * Opening a Feature Request to add compatibility to Windows → [ivadomed/ivadomed#604](https://github.com/ivadomed/ivadomed/issues/604)
> * Adding  explanation for Windows incompatibility in installation docs → [ivadomed/ivadomed#605](https://github.com/ivadomed/ivadomed/pull/605)
> * Specify Python version in setup.py → [ivadomed/ivadomed#603](https://github.com/ivadomed/ivadomed/pull/603)
>
> > In your table you include some examples of registration packages that use deep learning, yet ivadomed does not yet cover these functionalities (yet (?)). Are you planning on covering this use case in the future? Additionally, would you consider explaining this in your paper, and adding [DeepReg](https://doi.org/10.21105/joss.02705) to your table/citing it?
>
> Thank you for this relevant suggestion! Indeed, `ivadomed` does not cover registration (yet). We have added a column for `DeepReg`, see [PR618](https://github.com/ivadomed/ivadomed/pull/618). Please feel free to comment/suggest modifications if we missed some features of `DeepReg` (see preview [here](https://ivado-medical-imaging--618.org.readthedocs.build/en/618/purpose.html)).

Hi @charleygros, thanks for the response, the reference for DeepReg is accurate in your able. Now just waiting on resolution of [the docs update re Windows](https://github.com/ivadomed/ivadomed/pull/605), then all good to go from me 😄
