### 2021-01-11 1c6d40d
config_file.rst: final_activation field
