### 2021-01-11 e549161
config_file.rst: is_2d field
