### 2021-01-11 6502c1a
config_file.rst: fix bn discrepancy
