### 2021-01-11 de758d7
Update docs/source/configuration_file.rst
Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>
