### 2021-01-11 77150bb
configuration_file.rst: add info on bn_momentum
