### 2021-01-11 d530746
Add explanation for Windows incompatibility in installation docs (#605)
Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>
