### 2021-01-11 e67c84a
Revert "Add explanation for Windows incompatibility in installation docs (#605)"
This reverts commit d5307469affbb59c1a22f3835ff91d4e517b895d.
