### 2021-01-11 b44e063
remove colon from supported OS heading
