### 2021-01-11 7916f91
update supported OS paragraph to include WSL
