### 2021-01-11 6ed0052
add call to imed_utils argparser in extract_small_dataset
