### 2021-01-11 b51a098
fix contrast_list bug in extract_small_dataset, clean up code
