+++ ivadomed/scripts/extract_small_dataset.py
@@ -20,10 +20,12 @@ def get_parser():
     parser.add_argument("-o", "--output", required=True,
                         help="Output BIDS Folder.")
     parser.add_argument("-s", "--seed", required=False, default=-1,
-                        help="Set np.random.RandomState to ensure reproducibility: the same subjects will be selected "
-                             "if the script is run several times on the same dataset. Set to -1 (default) otherwise.")
+                        help="""Set np.random.RandomState to ensure reproducibility:
+                                the same subjects will be selected f the script is run several
+                                times on the same dataset. Set to -1 (default) otherwise.""")
     parser.add_argument("-d", "--derivatives", required=False, default=1,
-                        help="1: Include derivatives/labels content. 0: Do not include derivatives/labels content.")
+                        help="""1: Include derivatives/labels content.
+                                0: Do not include derivatives/labels content.""")
     return parser
 
 
@@ -47,21 +49,22 @@ def extract_small_dataset(input, output, n=10, contrast_list=None, include_deriv
 
     Example::
 
-         ivadomed_extract_small_dataset -i path/to/BIDS/dataset -o path/of/small/BIDS/dataset -n 10 -c T1w,T2w -d 0 -s 1234
+         ivadomed_extract_small_dataset -i path/to/BIDS/dataset -o path/of/small/BIDS/dataset \
+            -n 10 -c T1w,T2w -d 0 -s 1234
 
     Args:
         input (str): Input BIDS folder. Flag: ``--input``, ``-i``
         output (str): Output folder. Flag: ``--output``, ``-o``
         n (int): Number of subjects in the output folder. Flag: ``--number``, ``-n``
-        contrast_list (list): List of image contrasts to include. If set to None, then all available contrasts are
-            included. Flag: ``--contrasts``, ``-c``
-        include_derivatives (bool): If True, derivatives/labels/ content is also copied, only the raw images otherwise.
-                                    Flag: ``--derivatives``, ``-d``
-        seed (int): Set np.random.RandomState to ensure reproducibility: the same subjects will be selected if the
-            function is run several times on the same dataset. If set to -1, each function run is independent.
-            Flag: ``--seed``, ``-s``.
+        contrast_list (list): List of image contrasts to include. If set to None, then all
+            available contrasts are included. Flag: ``--contrasts``, ``-c``
+        include_derivatives (bool): If True, derivatives/labels/ content is also copied,
+            only the raw images otherwise. Flag: ``--derivatives``, ``-d``
+        seed (int): Set np.random.RandomState to ensure reproducibility: the same subjects will be
+            selected if the function is run several times on the same dataset. If set to -1,
+            each function run is independent. Flag: ``--seed``, ``-s``.
     """
-    # Create o folders
+    # Create output folders
     if not os.path.isdir(output):
         os.makedirs(output)
     if include_derivatives:
@@ -110,7 +113,8 @@ def extract_small_dataset(input, output, n=10, contrast_list=None, include_deriv
     if contrast_list:
         remove_some_contrasts(output, subject_random_list, contrast_list)
         if include_derivatives:
-            remove_some_contrasts(os.path.join(output, "derivatives", "labels"), subject_random_list, contrast_list)
+            remove_some_contrasts(os.path.join(output, "derivatives", "labels"),
+                                  subject_random_list, contrast_list)
 
     # Copy dataset_description.json
     idatasetjson = os.path.join(input, "dataset_description.json")
@@ -134,7 +138,11 @@ def main(args=None):
     imed_utils.init_ivadomed()
     parser = get_parser()
     args = imed_utils.get_arguments(parser, args)
-    extract_small_dataset(args.input, args.output, int(args.number), args.contrasts.split(","),
[…]
