### 2021-01-11 37b2a13
add functional tests for extract_small_dataset
