### 2021-01-11 c8192e3
add call to imed_utils argparser in prepare_dataset_vertebral_labelling
