+++ ivadomed/scripts/prepare_dataset_vertebral_labeling.py
@@ -50,7 +50,7 @@ def extract_mid_slice_and_convert_coordinates_to_heatmaps(path, suffix, aim=-1):
     Example::
 
         ivadomed_prepare_dataset_vertebral_labeling -p path/to/bids -s _T2w -a 0
-    
+
     Args:
         path (string): path to BIDS dataset form which images will be generated. Flag: ``--path``, ``-p``
         suffix (string): suffix of image that will be processed (e.g., T2w). Flag: ``--suffix``, ``-s``
@@ -103,15 +103,13 @@ def get_parser():
     return parser
 
 
-def main():
+def main(args=None):
     imed_utils.init_ivadomed()
-
     parser = get_parser()
-    args = parser.parse_args()
+    args = imed_utils.get_arguments(parser, args)
     bids_path = args.path
     suffix = args.suffix
     aim = args.aim
-    # Run Script
     extract_mid_slice_and_convert_coordinates_to_heatmaps(bids_path, suffix, aim)
 
 if __name__=='__main__':
