### 2021-01-11 9a26412
add test for prepare_dataset_vertebral_labelling, code cleanup
