### 2021-01-11 1644065
add call to imed_utils argparser in training_curve, clean up code
