+++ ivadomed/scripts/training_curve.py
@@ -8,25 +8,26 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from textwrap import wrap
 from tensorboard.backend.event_processing.event_accumulator import EventAccumulator
-
-from ivadomed.utils import init_ivadomed
+from ivadomed import utils as imed_utils
 
 
 def get_parser():
     parser = argparse.ArgumentParser()
     parser.add_argument("-i", "--input", required=True, type=str,
-                        help="Input log directory. If using --multiple, this parameter indicates the suffix path of all"
-                             " log directories of interest. To compare trainings or set of trainings (using "
-                             "``--multiple``) with subplots, please list the paths by separating them with commas, eg "
-                             "path_log_dir1,path_logdir2.")
+                        help="""Input log directory. If using --multiple, this parameter indicates
+                                the suffix path of all log directories of interest. To compare
+                                trainings or set of trainings (using ``--multiple``) with subplots,
+                                please list the paths by separating them with commas, e.g.
+                                path_log_dir1,path_logdir2.""")
     parser.add_argument("--multiple", required=False, dest="multiple", action='store_true',
-                        help="Multiple log directories are considered: all available folders with -i as "
-                             "prefix. The plot represents the mean value (hard line) surrounded by the standard "
-                             "deviation envelope.")
+                        help="""Multiple log directories are considered: all available folders
+                                with -i as prefix. The plot represents the mean value (hard line)
+                                surrounded by the standard deviation envelope.""")
     parser.add_argument("-y", "--ylim_loss", required=False, type=str,
-                        help="Indicates the limits on the y-axis for the loss plots, otherwise these limits are "
-                             "automatically defined. Please separate the lower and the upper limit by a comma, eg -1,0."
-                             "Note: for the validation metrics: the y-limits are always 0.0 and 1.0.")
+                        help="""Indicates the limits on the y-axis for the loss plots, otherwise
+                                these limits are automatically defined. Please separate the lower
+                                and the upper limit by a comma, eg -1,0.
+                                Note: for the validation metrics: the y-limits are always 0.0 and 1.0.""")
     parser.add_argument("-o", "--output", required=True, type=str,
                         help="Output folder.")
     return parser
@@ -96,29 +97,30 @@ def run_plot_training_curves(input_folder, output_folder, multiple_training=Fals
         :width: 600px
         :align: center
 
-    ... or multiple (using ``multiple_training=True``). In that case, the hard line represents the mean value across the
-    trainings whereas the envelope represents the standard deviation:
+    ... or multiple (using ``multiple_training=True``). In that case, the hard line represents
+    the mean value across the trainings whereas the envelope represents the standard deviation:
 
     .. image:: https://raw.githubusercontent.com/ivadomed/doc-figures/main/scripts/plot_loss_multiple.png
         :width: 600px
         :align: center
 
-    It is also possible to compare multiple trainings (or set of trainings) by listing them in ``-i``, separeted by
-    commas:
+    It is also possible to compare multiple trainings (or set of trainings) by listing them
+    in ``-i``, separated by commas:
 
     .. image:: https://raw.githubusercontent.com/ivadomed/doc-figures/main/scripts/plot_loss_mosaic.png
         :width: 600px
         :align: center
 
     Args:
-        input_folder (str): Log directory name. Flag: ``--input``, ``-i``. If using ``--multiple``, this parameter indicates the
[…]
