### 2021-01-11 #2868 @charleygros
> Now just waiting on resolution of [the docs update re Windows](https://github.com/ivadomed/ivadomed/pull/605), then all good to go from me smile

Now merged, as well as an [additional PR](https://github.com/ivadomed/ivadomed/pull/621) about the same matter. Thanks again @NMontanaBrown :-)
