### 2021-01-11 #2868 @charleygros
Hi @lbugnon,

I noticed you have not checked off the "Performance" item:

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

May I ask if you have some feedback for us? Cheers
