### 2021-01-11 f0ad66f
training.py: imed_visualize.AnimatedGif
