### 2021-01-11 f09eb5d
loader.py: soft_gt only before network
