### 2021-01-11 c5217e3
loader.py: soft_gt=False as uint8
