### 2021-01-11 ff5df4c
doc: add info on packaged model
