### 2021-01-11 accb895
Add config extension
Co-authored-by: charleygros <charley.gros@gmail.com>
