### 2021-01-12 65ef6a1
update data_functional_testing release version
