+++ ivadomed/scripts/download_data.py
@@ -39,7 +39,7 @@ DICT_URL = {
         "url": ["https://github.com/ivadomed/model_find_disc_t2/archive/r20200928.zip"],
         "description": "Intervertebral disc detection model trained on T2-weighted images."},
     "data_functional_testing": {
-        "url": ["https://github.com/ivadomed/data_functional_testing/archive/r20201218.zip"],
+        "url": ["https://github.com/ivadomed/data_functional_testing/archive/r20210111.zip"],
         "description": "Data used for functional testing in Ivadomed."
     }
 
