### 2021-01-12 8f8885a
add test for training_curve, clean code
