### 2021-01-12 ce6cb78
add call to imed_utils argparser in visualize_transforms, clean up code
