### 2021-01-12 136e923
add test for visualize_transforms
