### 2021-01-12 e9bae09
Add section for command line tools
