+++ docs/source/usage.rst
@@ -1,17 +1,0 @@
-Usage
-=====
-
-Command line tools
-------------------
-
-New model can be generated using the command-line tool from the
-terminal:
-
-::
-
-    ivadomed -c path/to/config_file.json
-
-where ``config_file.json`` is a configuration file, which parameters are
-described in the :ref:`configuration_file:Configuration File`.
-
-Please see section ``TUTORIALS`` to run this command on an example dataset.
