### 2021-01-12 88981a7
model_name
Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>
