### 2021-01-12 d5264d1
Change model example
Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>
