+++ docs/source/pretrained_models.rst
@@ -12,7 +12,7 @@ For convenience, the following pre-trained models are ready-to-use:
 - `model_find_disc_t2 <https://github.com/ivadomed/model_find_disc_t2/archive/r20200928.zip>`_: Intervertebral disc detection model trained on T2-weighted images.
 
 Packaged model format
-=====================
+---------------------
 Each folder contains a model (.pt or .onnx) with its corresponding configuration file (.json). The packaged model is
 automatically generated during training. The folder containing the packaged model will be saved at the path specified by
 the key ``log_directory`` in the configuration file. The packaged model, the configuration file, and the model file will
