### 2021-01-12 1259c42
Remove background class from HeMIS prediction output
