### 2021-01-12 ceed2ba
Rename filter to filter_absent_class, update docs
