### 2021-01-12 c9540fb
Change functional tests so that data isn't downloaded every time, but copied
