+++ testing/functional_tests/test_extract_small_dataset.py
@@ -1,19 +1,17 @@
 import logging
 import os
-from cli_base import remove_tmp_dir, __tmp_dir__, create_tmp_dir, download_dataset
+from cli_base import remove_tmp_dir, __tmp_dir__, create_tmp_dir, __data_testing_dir__
 from ivadomed.scripts import extract_small_dataset
 logger = logging.getLogger(__name__)
 
 
 def setup_function():
     create_tmp_dir()
-    download_dataset("data_functional_testing")
 
 
 def test_extract_small_dataset_default_n():
-    __input_dir__ = os.path.join(__tmp_dir__, 'data_functional_testing')
     __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset')
-    extract_small_dataset.main(args=['--input', __input_dir__,
+    extract_small_dataset.main(args=['--input', __data_testing_dir__,
                                      '--output', __output_dir__])
     assert os.path.exists(__output_dir__)
     output_dir_list = os.listdir(__output_dir__)
@@ -26,9 +24,8 @@ def test_extract_small_dataset_default_n():
 
 
 def test_extract_small_dataset_n_2():
-    __input_dir__ = os.path.join(__tmp_dir__, 'data_functional_testing')
     __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_2')
-    extract_small_dataset.main(args=['--input', __input_dir__,
+    extract_small_dataset.main(args=['--input', __data_testing_dir__,
                                      '--output', __output_dir__,
                                      '-n', '2'])
     assert os.path.exists(__output_dir__)
@@ -44,9 +41,8 @@ def test_extract_small_dataset_n_2():
         'sub-unf03' not in output_dir_list
 
 def test_extract_small_dataset_no_derivatives():
-    __input_dir__ = os.path.join(__tmp_dir__, 'data_functional_testing')
     __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_3')
-    extract_small_dataset.main(args=['--input', __input_dir__,
+    extract_small_dataset.main(args=['--input', __data_testing_dir__,
                                      '--output', __output_dir__,
                                      '-d', '0'])
     assert os.path.exists(__output_dir__)
@@ -59,9 +55,8 @@ def test_extract_small_dataset_no_derivatives():
         'sub-unf03' in output_dir_list
 
 def test_extract_small_dataset_contrast_list():
-    __input_dir__ = os.path.join(__tmp_dir__, 'data_functional_testing')
     __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_4')
-    extract_small_dataset.main(args=['--input', __input_dir__,
+    extract_small_dataset.main(args=['--input', __data_testing_dir__,
                                      '--output', __output_dir__,
                                      '-c', 'T1w, T2w'])
     assert os.path.exists(__output_dir__)
