### 2021-01-12 4b584bc
move download of data_functional_testing to run_tests.yml
