### 2021-01-12 e5346c3
move data from fixtures folder to data_functional_testing repo
