### 2021-01-12 cccfd39
remove dirs_exist_ok as it isn't supported by python < 3.8
