### 2021-01-12 a454052
Update ivadomed/loader/utils.py
Co-authored-by: andreanne-lemay <49137243+andreanne-lemay@users.noreply.github.com>
