### 2021-01-12 f7ad2e3
update test_loader to use t_utils
