+++ testing/unit_tests/test_loader.py
@@ -1,14 +1,18 @@
 import os
-
 import pytest
-import pandas as pd
 import csv_diff
-
+import logging
+from t_utils import remove_tmp_dir, create_tmp_dir, __data_testing_dir__
 from ivadomed.loader import utils as imed_loader_utils
+logger = logging.getLogger(__name__)
+
+
+def setup_function():
+    create_tmp_dir()
 
 
 @pytest.mark.parametrize('loader_parameters', [{
-    "bids_path": "testing_data/microscopy_png",
+    "bids_path": os.path.join(__data_testing_dir__, "microscopy_png"),
     "bids_config": "ivadomed/config/config_bids.json",
     "target_suffix": [["_seg-myelin-manual", "_seg-axon-manual"]],
     "extensions": [".png"],
@@ -18,10 +22,13 @@ from ivadomed.loader import utils as imed_loader_utils
         "balance": {}
     }}])
 def test_bids_df_microscopy_png(loader_parameters):
-    # Test for microscopy png file format
-    # Test for _sessions.tsv and _scans.tsv files
-    # Test for target_suffix as a nested list
-    # Test for when no contrast_params are provided
+    """
+    Test for microscopy png file format
+    Test for _sessions.tsv and _scans.tsv files
+    Test for target_suffix as a nested list
+    Test for when no contrast_params are provided
+    """
+
     loader_params = loader_parameters
     loader_params["contrast_params"]["contrast_lst"] = loader_params["contrast_params"]["training_validation"]
     bids_path = loader_params["bids_path"]
@@ -37,7 +44,7 @@ def test_bids_df_microscopy_png(loader_parameters):
 
 
 @pytest.mark.parametrize('loader_parameters', [{
-    "bids_path": "testing_data",
+    "bids_path": __data_testing_dir__,
     "target_suffix": ["_seg-manual"],
     "extensions": [],
     "contrast_params": {
@@ -46,9 +53,11 @@ def test_bids_df_microscopy_png(loader_parameters):
         "balance": {}
     }}])
 def test_bids_df_anat(loader_parameters):
-    # Test for MRI anat nii.gz file format
-    # Test for when no file extensions are provided
-    # Test for multiple target_suffix
+    """
+    Test for MRI anat nii.gz file format
+    Test for when no file extensions are provided
+    Test for multiple target_suffix
+    """
     loader_params = loader_parameters
     loader_params["contrast_params"]["contrast_lst"] = loader_params["contrast_params"]["training_validation"]
     bids_path = loader_params["bids_path"]
@@ -60,4 +69,9 @@ def test_bids_df_anat(loader_parameters):
     csv_test = os.path.join(bids_path, "df_test.csv")
     df_test.to_csv(csv_test, index=False)
     diff = csv_diff.compare(csv_diff.load_csv(open(csv_ref)), csv_diff.load_csv(open(csv_test)))
-    assert diff == {'added': [], 'removed': [], 'changed': [], 'columns_added': [], 'columns_removed': []}
+    assert diff == {'added': [], 'removed': [], 'changed': [],
+                    'columns_added': [], 'columns_removed': []}
+
+
+def teardown_function():
+    remove_tmp_dir()
