### 2021-01-13 3a8f6bb
Add "split_output" param in split_dataset functions
