+++ ivadomed/loader/utils.py
@@ -96,7 +96,7 @@ def split_dataset(df, center_test_lst, split_method, random_seed, train_frac=0.8
     return X_train, X_val, X_test
 
 
-def split_dataset_new(df, data_testing, random_seed, train_frac=0.8, test_frac=0.1):
+def split_dataset_new(df, split_output, data_testing, random_seed, train_frac=0.8, test_frac=0.1):
     """Splits dataset into training, validation and testing datasets according to the data_type selected in data_testing.
     Example: If data_type is "institution_id", the centers associated to the subjects are split according the train, test and
     validation fraction whereas if data_type is "subject", the patients are directly separated according to these fractions.
@@ -113,22 +113,22 @@ def split_dataset_new(df, data_testing, random_seed, train_frac=0.8, test_frac=0
     # Init output lists
 
     # Get data_type and data_value from split parameters
-    data_type = data_testing['data_type']
+    data_type = data_testing['data_type'] if data_testing['data_type'] else split_output
     data_value = data_testing['data_value']
 
     # Filter dataframe with rows where data_type is not NAN
     df = df[df[data_type].notna()]
 
     # List unique data_type
-    data = sorted(df[data_type].unique().tolist())
+    data = sorted(df[split_output].unique().tolist())
 
     # Split test dataset according to data_value
     # Make sure that data_type coming from data_value are unseen during training
     if len(data_value) == 0 and test_frac != 0:
         test_frac = test_frac if test_frac >= 1 / len(data) else 1 / len(data)
         data_value, _ = train_test_split(data, train_size=test_frac, random_state=random_seed)
-    X_test = df[df[data_type].isin(data_value)][data_type].unique().tolist()
-    X_remain = df[~df[data_type].isin(data_value)][data_type].unique().tolist()
+    X_test = df[df[data_type].isin(data_value)][split_output].unique().tolist()
+    X_remain = df[~df[data_type].isin(data_value)][split_output].unique().tolist()
 
     # Split remainder in train/valid or train/valid/test datasets using sklearn function
     train_frac = train_frac*len(data)/len(X_remain)
@@ -213,7 +213,7 @@ def get_new_subject_split(path_folder, center_test, split_method, random_seed,
     return train_lst, valid_lst, test_lst
 
 
-def get_new_subject_split_new(df, data_testing, random_seed,
+def get_new_subject_split_new(df, split_output, data_testing, random_seed,
                           train_frac, test_frac, log_directory, balance, subject_selection=None):
     """Randomly split dataset between training / validation / testing.
 
@@ -261,6 +261,7 @@ def get_new_subject_split_new(df, data_testing, random_seed,
     for df_tmp in df_list:
         # Split dataset on each section of subjects
         train_tmp, valid_tmp, test_tmp = split_dataset_new(df=df_tmp,
+                                                        split_output=split_output,
                                                         data_testing=data_testing,
                                                         random_seed=random_seed,
                                                         train_frac=train_frac,
@@ -326,6 +327,7 @@ def get_subdatasets_subjects_list_new(split_params, df, log_directory, subject_s
         train_lst, valid_lst, test_lst = old_split['train'], old_split['valid'], old_split['test']
     else:
         train_lst, valid_lst, test_lst = get_new_subject_split_new(df=df,
+                                                               split_output=split_params['split_output'],
                                                                data_testing=split_params['data_testing'],
                                                                random_seed=split_params['random_seed'],
                                                                train_frac=split_params['train_fraction'],
