### 2021-01-13 9f4c97b
add docstrings, close pools on failure so that pytest doesn't hang in test_automate_training
