### 2021-01-13 de348cc
Update docstrings
