### 2021-01-13 49457ec
remove unit_base functionality, will address in separate pull request
