### 2021-01-13 5a55940
remove mytest hdf5 file
