### 2021-01-13 7e007c1
try moving set_start_method to initial if name == main
