+++ ivadomed/scripts/automate_training.py
@@ -285,9 +285,6 @@ def automate_training(config, param, fixed_split, all_combin, n_iterations=1, ru
         os.makedirs(output_dir)
     if not output_dir:
         output_dir = ""
-    # CUDA problem when forking process
-    # https://github.com/pytorch/pytorch/issues/2517
-    mp.set_start_method('spawn')
 
     # Run all configs on a separate process, with a maximum of n_gpus  processes at a given time
     pool = mp.Pool(processes=len(initial_config["gpu"]))
@@ -410,4 +407,7 @@ def main(args=None):
 
 
 if __name__ == '__main__':
+    # CUDA problem when forking process
+    # https://github.com/pytorch/pytorch/issues/2517
+    mp.set_start_method('spawn')
     main()
