### 2021-01-14 5f2d6fe
Fix scipy ndimage import issue for newer scipy version
