### 2021-01-14 3392fc1
clean utils, use collections.abc due to deprecation
