+++ ivadomed/loader/utils.py
@@ -1,8 +1,7 @@
-import collections
+import collections.abc
 import re
 import os
 import logging
-
 import numpy as np
 import pandas as pd
 import torch
@@ -12,8 +11,7 @@ from sklearn.model_selection import train_test_split
 from torch._six import string_classes, int_classes
 from ivadomed import utils as imed_utils
 import nibabel as nib
-import bids as pybids   #"bids" is already taken by bids_neuropoly
-import pandas as pd
+import bids as pybids   # "bids" is already taken by bids_neuropoly
 import itertools
 
 __numpy_type_map = {
@@ -27,7 +25,8 @@ __numpy_type_map = {
     'uint8': torch.ByteTensor,
 }
 
-TRANSFORM_PARAMS = ['elastic', 'rotation', 'scale', 'offset', 'crop_params', 'reverse', 'translation', 'gaussian_noise']
+TRANSFORM_PARAMS = ['elastic', 'rotation', 'scale', 'offset', 'crop_params', 'reverse',
+                    'translation', 'gaussian_noise']
 
 logger = logging.getLogger(__name__)
 
@@ -110,8 +109,9 @@ def get_new_subject_split(path_folder, center_test, split_method, random_seed,
         train_frac (float): Training dataset proportion, between 0 and 1.
         test_frac (float): Testing dataset proportionm between 0 and 1.
         log_directory (string): Output folder.
-        balance (string): Metadata contained in "participants.tsv" file with categorical values. Each category will be
-        evenly distributed in the training, validation and testing datasets.
+        balance (string): Metadata contained in "participants.tsv" file with categorical values.
+            Each category will be evenly distributed in the training, validation and testing
+            datasets.
         subject_selection (dict): Used to specify a custom subject selection from a dataset.
 
     Returns:
@@ -136,8 +136,8 @@ def get_new_subject_split(path_folder, center_test, split_method, random_seed,
         if balance in df.keys():
             df_list = [df[df[balance] == k] for k in df[balance].unique().tolist()]
         else:
-            logger.warning("No column named '{}' was found in 'participants.tsv' file. Not taken into account to split "
-                           "the dataset.".format(balance))
+            logger.warning(f"""No column named '{balance}' was found in 'participants.tsv' file.
+                               Not taken into account to split the dataset.""")
             df_list = [df]
     else:
         df_list = [df]
@@ -225,9 +225,9 @@ def imed_collate(batch):
         return torch.DoubleTensor(batch)
     elif isinstance(batch[0], string_classes):
         return batch
-    elif isinstance(batch[0], collections.Mapping):
+    elif isinstance(batch[0], collections.abc.Mapping):
         return {key: imed_collate([d[key] for d in batch]) for key in batch[0]}
-    elif isinstance(batch[0], collections.Sequence):
+    elif isinstance(batch[0], collections.abc.Sequence):
         return [imed_collate(samples) for samples in batch]
 
     return batch
@@ -236,8 +236,9 @@ def imed_collate(batch):
 def filter_roi(roi_data, nb_nonzero_thr):
     """Filter slices from dataset using ROI data.
 
-    This function filters slices (roi_data) where the number of non-zero voxels within the ROI slice (e.g. centerline,
-    SC segmentation) is inferior or equal to a given threshold (nb_nonzero_thr).
+    This function filters slices (roi_data) where the number of non-zero voxels within the
+    ROI slice (e.g. centerline, SC segmentation) is inferior or equal to a given threshold
+    (nb_nonzero_thr).
 
     Args:
         roi_data (nd.array): ROI slice.
@@ -255,7 +256,8 @@ def orient_img_hwd(data, slice_axis):
 
     Args:
         data (ndarray): RAS oriented data.
[…]
