### 2021-01-14 8585db9
update test_sampler to use t_utils
