### 2021-01-16 #2868 @ssinad
> > Hey @charleygros,
> > I really like your software. However, I have some small comments:
> >
> > * **Installation:** With Python 3.9 it does not work, although the documentation states, it should just be Python>=3.6
> >
> > ```shell
> > pip install ivadomed
> > Collecting ivadomed
> >   Downloading ivadomed-2.6.0-py3-none-any.whl (139 kB)
> >      |████████████████████████████████| 139 kB 8.9 MB/s
> >   Downloading ivadomed-2.5.0-py3-none-any.whl (137 kB)
> >      |████████████████████████████████| 137 kB 17.9 MB/s
> >   Downloading ivadomed-2.4.0-py3-none-any.whl (132 kB)
> >      |████████████████████████████████| 132 kB 23.7 MB/s
> >   Downloading ivadomed-2.3.1-py3-none-any.whl (129 kB)
> >      |████████████████████████████████| 129 kB 24.1 MB/s
> >   Downloading ivadomed-2.2.1-py3-none-any.whl (127 kB)
> >      |████████████████████████████████| 127 kB 28.2 MB/s
> >   Downloading ivadomed-2.2.0-py3-none-any.whl (127 kB)
> >      |████████████████████████████████| 127 kB 23.9 MB/s
> >   Downloading ivadomed-2.1.0-py3-none-any.whl (110 kB)
> >      |████████████████████████████████| 110 kB 30.8 MB/s
> >   Downloading ivadomed-2.0.2-py3-none-any.whl (108 kB)
> >      |████████████████████████████████| 108 kB 29.9 MB/s
> >   Downloading ivadomed-2.0.1-py3-none-any.whl (108 kB)
> >      |████████████████████████████████| 108 kB 30.9 MB/s
> >   Downloading ivadomed-1.2.1-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 1.8 MB/s
> >   Downloading ivadomed-1.2-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 5.0 MB/s
> >   Downloading ivadomed-1.1-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 4.6 MB/s
> >   Downloading ivadomed-1.0-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 2.4 MB/s
> >   Downloading ivadomed-0.14-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 4.8 MB/s
> >   Downloading ivadomed-0.13-py3-none-any.whl (68 kB)
> >      |████████████████████████████████| 68 kB 457 kB/s
> >   Downloading ivadomed-0.12-py3-none-any.whl (66 kB)
> >      |████████████████████████████████| 66 kB 3.5 MB/s
> >   Downloading ivadomed-0.11-py3-none-any.whl (66 kB)
> >      |████████████████████████████████| 66 kB 3.3 MB/s
> > ERROR: Cannot install ivadomed==0.11, ivadomed==0.12, ivadomed==0.13, ivadomed==0.14, ivadomed==1.0, ivadomed==1.1, ivadomed==1.2, ivadomed==1.2.1, ivadomed==2.0.1, ivadomed==2.0.2, ivadomed==2.1.0, ivadomed==2.2.0, ivadomed==2.2.1, ivadomed==2.3.1, ivadomed==2.4.0, ivadomed==2.5.0 and ivadomed==2.6.0 because these package versions have conflicting dependencies.
> >
> > The conflict is caused by:
> >     ivadomed 2.6.0 depends on onnxruntime==1.4.0
> >     ivadomed 2.5.0 depends on onnxruntime==1.4.0
> >     ivadomed 2.4.0 depends on onnxruntime
> >     ivadomed 2.3.1 depends on torchvision>=0.6.0
> >     ivadomed 2.2.1 depends on torchvision>=0.6.0
> >     ivadomed 2.2.0 depends on torchvision>=0.6.0
> >     ivadomed 2.1.0 depends on torchvision>=0.6.0
> >     ivadomed 2.0.2 depends on torchvision>=0.6.0
> >     ivadomed 2.0.1 depends on torchvision>=0.6.0
> >     ivadomed 1.2.1 depends on torchvision>=0.6.0
> >     ivadomed 1.2 depends on torchvision>=0.6.0
> >     ivadomed 1.1 depends on torchvision>=0.6.0
> >     ivadomed 1.0 depends on torchvision>=0.6.0
> >     ivadomed 0.14 depends on torchvision>=0.6.0
> >     ivadomed 0.13 depends on torchvision>=0.6.0
> >     ivadomed 0.12 depends on torchvision>=0.6.0
> >     ivadomed 0.11 depends on torchvision>=0.6.0
> >
> > To fix this you could try to:
> > 1. loosen the range of package versions you've specified
> > 2. remove package versions to allow pip attempt to solve the dependency conflict
> >
> > ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
> > ```
> >
> >
> > With Python 3.8.5 it works fine! Maybe having the installation instructions also in the readme and not only in the linked documentation would be nice.
> >
> > * **State of Need:**
> >   While it is clearly stated, why the software is necessary and how it compares to other available software, the table at https://ivadomed.org/en/latest/purpose.html#comparison-with-other-projects is not nice to look at. Maybe divide it into several tables with major parts.
> > * **Automated tests**: While tests are present and coverage is reported correctly, it would be nice having a badge pointing to the actual outcome of GitHub actions.
> > * **Community Guidelines**: While they are available at https://ivadomed.org/en/latest/contributing.html and in the repository wiki, it is not the usual place for GitHub repositories. SO it would be nice to also have a file like `CODE_OF_CONDUCT.md` (which must only point to the available resources) as well as issue and Pull-Request templates pointing to the relevant parts of the wiki.
>
> Hi 👋 @charleygros, thanks for the work.
>
> # Installation
> Currently I am having the same problems with installation as @lbugnon reported earlier, with differing outcomes depending on OS.
> MacOS installation proceeds correctly for latest release.
> Ubuntu 18.04.3 installation proceeds correctly for latest release.
> Windows OS left me with the same error as above, for some reason.
>
> ```
> Collecting ivadomed
>   Using cached ivadomed-2.6.1-py3-none-any.whl (145 kB)
>   Using cached ivadomed-2.6.0-py3-none-any.whl (139 kB)
>   Using cached ivadomed-2.5.0-py3-none-any.whl (137 kB)
>   Using cached ivadomed-2.4.0-py3-none-any.whl (132 kB)
>   Using cached ivadomed-2.3.1-py3-none-any.whl (129 kB)
>   Using cached ivadomed-2.2.1-py3-none-any.whl (127 kB)
>   Using cached ivadomed-2.2.0-py3-none-any.whl (127 kB)
>   Using cached ivadomed-2.1.0-py3-none-any.whl (110 kB)
>   Using cached ivadomed-2.0.2-py3-none-any.whl (108 kB)
>   Using cached ivadomed-2.0.1-py3-none-any.whl (108 kB)
>   Using cached ivadomed-1.2.1-py3-none-any.whl (68 kB)
>   Using cached ivadomed-1.2-py3-none-any.whl (68 kB)
>   Using cached ivadomed-1.1-py3-none-any.whl (68 kB)
>   Using cached ivadomed-1.0-py3-none-any.whl (68 kB)
>   Using cached ivadomed-0.14-py3-none-any.whl (68 kB)
>   Using cached ivadomed-0.13-py3-none-any.whl (68 kB)
>   Using cached ivadomed-0.12-py3-none-any.whl (66 kB)
>   Using cached ivadomed-0.11-py3-none-any.whl (66 kB)
> ERROR: Cannot install ivadomed==0.11, ivadomed==0.12, ivadomed==0.13, ivadomed==0.14, ivadomed==1.0, ivadomed==1.1, ivadomed==1.2, ivadomed==1.2.1, ivadomed==2.0.1, ivadomed==2.0.2, ivadomed==2.1.0, ivadomed==2.2.0, ivadomed==2.2.1, ivadomed==2.3.1, ivadomed==2.4.0, ivadomed==2.5.0, ivadomed==2.6.0 and ivadomed==2.6.1 because these package versions have conflicting dependencies.
>
> The conflict is caused by:
>     ivadomed 2.6.1 depends on torchvision==0.6.0
>     ivadomed 2.6.0 depends on torchvision==0.6.0
>     ivadomed 2.5.0 depends on torchvision==0.6.0
>     ivadomed 2.4.0 depends on torchvision==0.6.0
>     ivadomed 2.3.1 depends on torchvision>=0.6.0
>     ivadomed 2.2.1 depends on torchvision>=0.6.0
>     ivadomed 2.2.0 depends on torchvision>=0.6.0
>     ivadomed 2.1.0 depends on torchvision>=0.6.0
>     ivadomed 2.0.2 depends on torchvision>=0.6.0
>     ivadomed 2.0.1 depends on torchvision>=0.6.0
>     ivadomed 1.2.1 depends on torchvision>=0.6.0
>     ivadomed 1.2 depends on torchvision>=0.6.0
>     ivadomed 1.1 depends on torchvision>=0.6.0
>     ivadomed 1.0 depends on torchvision>=0.6.0
>     ivadomed 0.14 depends on torchvision>=0.6.0
>     ivadomed 0.13 depends on torchvision>=0.6.0
>     ivadomed 0.12 depends on torchvision>=0.6.0
>     ivadomed 0.11 depends on torchvision>=0.6.0
>
> To fix this you could try to:
> 1. loosen the range of package versions you've specified
> 2. remove package versions to allow pip attempt to solve the dependency conflict
>
> ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
> ```
>
> Specifically, the error seems to be with torchvision:
>
> ```
> Collecting ivadomed==2.6.1
>   Using cached ivadomed-2.6.1-py3-none-any.whl (145 kB)
> ERROR: Could not find a version that satisfies the requirement torchvision==0.6.0 (from ivadomed)
> ERROR: No matching distribution found for torchvision==0.6.0
> ```
>
> I tried this with various python versions to no avail. Overall, this seems to be related to WindowsOS, as the installation proceeds fine on other OS. I would suggest updating your RTD installation section and the README to state installation works as described for MacOS and Linux, but not currently supported for Windows. It may be something you want to fix or not - but this will clarify upfront to users that installation works as intended except on Windows and avoid head scratching at the first stage.
>
> # Statement of Need
>
> There is a table in the paper linking to a few of the packages in the scope of medical imaging and deep learning - as you note the field rapidly evolves and keeping track of all the packages is tough, so this table is useful. However, could I suggest to change the last sentence of the paper to highlight the completeness of your package in comparison to the others?
>
> In your table you include some examples of registration packages that use deep learning, yet ivadomed does not yet cover these functionalities (yet (?)). Are you planning on covering this use case in the future? Additionally, would you consider explaining this in your paper, and adding [DeepReg](https://doi.org/10.21105/joss.02705) to your table/citing it?

@NMontanaBrown  Can you verify the version of Python used on each of these operating systems?
