### 2021-01-16 #2868 @NMontanaBrown
@ssinad

I used conda for installation. So, on all OS, do:

`conda create --name ivadomed python=3.x`
`pip install ivadomed`

so ivadomed latest: 2.6.1

- MacOS: 3.7.1, 3.6, 3.8. 3.9:  downloads all the wheels for ivadomed all down to 1.1, and installs the 1.1 package.
- Ubuntu 18.04.3: 3.6, 3.7.9, 3.8.5, 3.9: downloads the wheels as with macOS - installs ivadomed 1.1.
- Windows OS: None, from 3.6 - 3.9.
