### 2021-01-16 #2868 @NMontanaBrown
> @NMontanaBrown
> Can you try this one instead?
>
> ```
> pip install ivadomed -f https://download.pytorch.org/whl/torch_stable.html
> ```

With which OS?
