### 2021-01-16 #2868 @ssinad
> > @NMontanaBrown
> > Can you try this one instead?
> > ```
> > pip install ivadomed -f https://download.pytorch.org/whl/torch_stable.html
> > ```
>
> With which OS?

@NMontanaBrown
Preferably Windows, but feel free to test it with others too.
