### 2021-01-19 #2868 @NMontanaBrown
> > > @NMontanaBrown
> > > Can you try this one instead?
> > > ```
> > > pip install ivadomed -f https://download.pytorch.org/whl/torch_stable.html
> > > ```
> >
> >
> > With which OS?
>
> @NMontanaBrown
> Preferably Windows, but feel free to test it with others too.

Apologies, this slipped my mind.

I have tried this with Windows, it seems to fix the problem, install proceeds correctly on Windows! 🍾 😄

I have not tried it with Linux or MacOS as these install fine as is.
