### 2021-02-09 #2868 @cMadan
@NMontanaBrown @justusschock @lbugnon, thank you for the thorough reviews, they are much appreciated!

@charleygros, everything looks good to me!

To move forward with accepting your submission, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
