### 2021-02-09 #2868 @charleygros
Many thanks for your help @cMadan !

Release of the software done --> [v2.7.0](https://github.com/ivadomed/ivadomed/tree/v2.7.0)
