### 2021-02-11 #2868 @jcohenadad
i noticed a broken link in the generated pdf. Page 1, the hyperlink under "each has their pros & cons." brings me there:
![image](https://user-images.githubusercontent.com/2482071/107651948-0dd14700-6c4e-11eb-843f-e98941d6de8f.png)
