### 2021-02-11 #2868 @jcohenadad
Same issue for hyperlink under "this example tutorial" in page 3:

![image](https://user-images.githubusercontent.com/2482071/107652375-82a48100-6c4e-11eb-82d8-facc42529c20.png)

The issue is possibly related to the "#" character in the URL (interpreted as %23)?
