### 2021-02-11 #2868 @jcohenadad
about the broken links mentioned above: when opening the generated PDF with acrobat reader, the links work. With macOs preview, i get those errors.
