### 2021-02-11 #2868 @danielskatz
For what it's worth, they work for me in Safari on macOS
