### 2021-02-11 #2868 @jcohenadad
@cMadan have you had similar issues with generated PDFs in the past? What do you suggest we can do? (e.g., remove the "#" in the URLs, but it would be a less direct redirection 🤷‍♂️)
