### 2021-02-11 #2868 @jcohenadad
> For what it's worth, they work for me in Safari on macOS

@danielskatz It doesn't work for me on Safari (see below). I'm running macOs 10.15.7 with Safari 14.0 and opened the PDF with Safari from my local station. I could not open it from the link in the [GitHub PR](https://github.com/openjournals/joss-papers/pull/2084/files?file-filters%5B%5D=.pdf#diff-c8d0a82b4ebde1bcd2aceb93d3cd0d9627c7b695af3899135d6f5cac8323fb4f).

![Screen Shot 2021-02-11 at 11 21 14 AM](https://user-images.githubusercontent.com/2482071/107665619-9e625400-6c5b-11eb-9cb5-d41608d5ef71.png)
