### 2021-02-11 #2868 @danielskatz
I'm on macOS 11.2 and Safari 14.0.3
