### 2021-02-11 #2868 @justusschock
Works for me on macOS 11.2 with both, safari and chrome
