### 2021-02-11 #2868 @jcohenadad
Just for reference, this is a known issue. In [this thread](https://github.com/latex3/hyperref/issues/110), they suggest to replace the hyperlinks with bitly links, but I don't know if we could rely on them in the long term.
