### 2021-02-11 #2868 @danielskatz
Given that this is resolved on current macOS, this appears to be a bug apple fixed, so I don't think a workaround is needed as we go forward.
