### 2021-02-11 #2868 @Drulex
In the genreated link, the fragment uri (ascii character `#`) is url-encoded (`%23`) which seems appropriate, but apparently not handled the same way by all browsers.

Maybe you can try tweaking whatever process generated the hyperlink to use `ascii` instead of url-encoded links? (warning this may break stuff, especially with hyperlinks that have embedded spaces)
