### 2021-02-12 #2868 @cMadan
fwiw. The links work correctly in Windows Chrome as well.
