### 2021-02-12 #2868 @jcohenadad
@NMontanaBrown, @justusschock, @lbugnon: many thanks for your reviews, @cMadan for editing the submission and @arfon for helping with everything. We had an awesome experience publishing with JOSS and we will certainly recommend it around! Keep up the good work! 🙏
