### 2021-05-06 #2868 @charleygros
@cMadan: Any idea how we can help Google Scholar to reference the JOSS paper? They found the arxiv version but not the JOSS one. Cheers
