### 2021-05-07 #2868 @arfon
@charleygros - as far as I am aware, Scholar _does_ index JOSS these days, but it's possible this happens less frequently than the arXiv?
